### 2025-10-13 #9228 @editorialbot
[PRE REVIEW]: DeepCausalMMM: A Deep Learning Framework for Marketing Mix Modeling with Causal Inference
**Submitting author:** @adityapt (<a href="http://orcid.org/0009-0008-9495-3932">Aditya Puttaparthi Tirumala</a>)
**Repository:** https://github.com/adityapt/deepcausalmmm/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.17
