### 2025-10-13 c79f792
Update paper.bib
Added DOI for Granger1969 - Investigating Causal Relations by Econometric Models and Cross-Spectral Methods.
