+++ JOSS/paper.bib
@@ -24,6 +24,7 @@
   volume       = {37},
   number       = {3},
   pages        = {424--438},
+  doi          = {10.2307/1912791},
   url          = {https://www.jstor.org/stable/1912791}
 }
 
