### 2025-10-15 b9a0ee4
Update paper.md
Make a few edits to the language.
