### 2025-10-15 e350c7b
Update paper.md
