### 2025-10-15 3acc91d
Update paper.md
