### 2025-10-17 28d665d
Update paper.md
update YAML config for processing
