+++ JOSS/paper.md
@@ -11,7 +11,10 @@ tags:
 authors:
   - name: "Aditya Puttaparthi Tirumala"
     orcid: "0009-0008-9495-3932"
-    affiliation: "Independent Researcher"
+    affiliation: "1"
+affiliations:
+  - index: 1
+    name: "Independent Researcher"    
 date: 5 October 2025
 bibliography: paper.bib
 archive_doi: 10.5281/zenodo.17274024
