### 2025-10-23 cc02da2
Release v1.0.18: Logging system overhaul, emoji removal, and documentation updates
Major Changes:
- Replaced all print() statements with logging in core library
- Removed all emojis from library code, changelog, and examples
- Updated JOSS paper with correct Meridian references and concept DOI
- Fixed misleading burn-in warning message in data processing
- Unified dashboard output to single dashboard_outputs folder
- Updated citation metadata to v1.0.18

Core Library:
- Added package-wide logging configuration in __init__.py
- Migrated train_model.py, data.py, unified_model.py to logging
- Mig
[…]
