+++ test_version.py
@@ -23,51 +23,51 @@ def run_command(cmd, cwd=None):
 
 def test_setuptools_scm():
     """Test setuptools-scm version generation."""
-    print("🔍 Testing setuptools-scm version generation...")
+    print(" Testing setuptools-scm version generation...")
     
     # Test 1: Check current version
     success, version, error = run_command("python -m setuptools_scm")
     if success:
-        print(f"✅ Current version: {version}")
+        print(f" Current version: {version}")
         if ".dev" in version:
-            print("⚠️  WARNING: Development version detected!")
+            print("  WARNING: Development version detected!")
             return False
     else:
-        print(f"❌ setuptools-scm failed: {error}")
+        print(f" setuptools-scm failed: {error}")
         return False
     
     return True
 
 def test_git_status():
     """Test Git repository status."""
-    print("\n🔍 Checking Git repository status...")
+    print("\n Checking Git repository status...")
     
     # Check for uncommitted changes
     success, output, error = run_command("git status --porcelain")
     if success:
         if output.strip():
-            print("⚠️  WARNING: Uncommitted changes detected:")
+            print("  WARNING: Uncommitted changes detected:")
             for line in output.strip().split('\n'):
                 print(f"    {line}")
             return False
         else:
-            print("✅ Working directory is clean")
+            print(" Working directory is clean")
     else:
-        print(f"❌ Git status check failed: {error}")
+        print(f" Git status check failed: {error}")
         return False
     
     return True
 
 def test_git_tags():
     """Test Git tags availability."""
-    print("\n🔍 Checking Git tags...")
+    print("\n Checking Git tags...")
     
     # Check if we're at a tag
     success, tag, error = run_command("git describe --exact-match --tags HEAD")
     if success:
-        print(f"✅ Currently at tag: {tag}")
+        print(f" Currently at tag: {tag}")
     else:
-        print("⚠️  Not at an exact tag")
+        print("  Not at an exact tag")
         
         # Show latest tag
         success, latest_tag, error = run_command("git describe --tags --abbrev=0")
@@ -83,7 +83,7 @@ def test_git_tags():
 
 def test_clean_build():
     """Test building in a clean temporary directory (simulates GitHub Actions)."""
-    print("\n🔍 Testing clean build (simulating GitHub Actions)...")
+    print("\n Testing clean build (simulating GitHub Actions)...")
     
     with tempfile.TemporaryDirectory() as temp_dir:
         temp_repo = Path(temp_dir) / "repo"
@@ -91,26 +91,26 @@ def test_clean_build():
         # Clone the current repository
         success, output, error = run_command(f"git clone . {temp_repo}")
         if not success:
-            print(f"❌ Failed to clone repository: {error}")
+            print(f" Failed to clone repository: {error}")
             return False
         
         # Test setuptools-scm in clean clone
         success, version, error = run_command("python -m setuptools_scm", cwd=temp_repo)
         if success:
-            print(f"✅ Clean clone version: {version}")
+            print(f" Clean clone version: {version}")
             if ".dev" in version:
-                print("⚠️  WARNING: Clean clone still shows development version!")
+                print("  WARNING: Clean clone still shows development version!")
                 print("    This suggests the issue is with Git history or tags")
                 return False
         else:
-            print(f"❌ setuptools-scm failed in clean clone: {error}")
+            print(f" setuptools-scm failed in clean clone: {error}")
             return False
     
     return True
 
 def test_build_package():
     """Test package building in isolated environment."""
-    print("\n🔍 Testing package build...")
+    print("\n Testing package build...")
     
     with tempfile.TemporaryDirectory() as temp_dir:
         temp_repo = Path(temp_dir) / "repo"
[…]
