+++ JOSS/paper.bib
@@ -39,6 +39,15 @@
   url          = {https://link.springer.com/book/10.1007/b109775}
 }
 
+@misc{Chan2017,
+  author       = {David Chan and Malcolm Perry},
+  title        = {Challenges and Opportunities in Media Mix Modeling},
+  year         = {2017},
+  howpublished = {Google Research White Paper},
+  note         = {accessed October 2025},
+  url          = {https://research.google/pubs/challenges-and-opportunities-in-media-mix-modeling/}
+}
+
 @article{Li2024Survey,
   author       = {Zongyu Li and Xiaobo Guo and Siwei Qiang},
   title        = {A Survey of Deep Causal Models and Their Industrial Applications},
@@ -49,6 +58,15 @@
   pages        = {14999--15023},
   doi          = {10.1007/s10462-024-10886-0},
   url          = {https://link.springer.com/article/10.1007/s10462-024-10886-0}
+}
+
+@software{Meridian2024,
+  author       = {Google Meridian Marketing Mix Modeling Team},
+  title        = {Meridian: Marketing Mix Modeling},
+  url          = {https://github.com/google/meridian},
+  version      = {1.2.1},
+  year         = {2025},
+  note         = {accessed October 2025}
 }
 
 @misc{LightweightMMM2022,
