+++ examples/dashboard_rmse_optimized.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-Beautiful RMSE-Optimized Dashboard with Config System
+Dashboard to visualize outputs
 ====================================================
 This implementation uses:
 - Config system for input parameters  
@@ -47,23 +47,23 @@ def get_viz_params(config):
 
 def load_config():
     """Load configuration from the single source of truth - config.py"""
-    print("📋 Loading Configuration...")
+    print(" Loading Configuration...")
     
     # Use only config.py - no dashboard overrides
     config = get_default_config()
     
-    print("   ✅ Configuration loaded from config.py")
-    print(f"   📊 Epochs: {config['n_epochs']}, Warm-start: {config['warm_start_epochs']}")
-    print(f"   🧠 Hidden units: {config['hidden_dim']}, Dropout: {config['dropout']}")
+    print("    Configuration loaded from config.py")
+    print(f"    Epochs: {config['n_epochs']}, Warm-start: {config['warm_start_epochs']}")
+    print(f"    Hidden units: {config['hidden_dim']}, Dropout: {config['dropout']}")
     return config
 
 def load_real_mmm_data(filepath="examples/data/MMM Data.csv"):
     """Load and process the real MMM Data.csv with robust missing data handling"""
-    print(f"📂 Loading Real MMM Data from: {filepath}")
+    print(f" Loading Real MMM Data from: {filepath}")
 
     try:
         df = pd.read_csv(filepath)
-        print(f"   📊 Loaded data shape: {df.shape}")
+        print(f"    Loaded data shape: {df.shape}")
 
         # Identify columns
         impression_cols = [col for col in df.columns if 'impressions_' in col]
@@ -97,8 +97,8 @@ def load_real_mmm_data(filepath="examples/data/MMM Data.csv"):
                 clean_name = clean_name.replace('_sm', '').replace('_', ' ').title()
             control_names.append(clean_name)
 
-        print(f"   📺 Media channels ({len(impression_cols)}): {media_names}")
-        print(f"   🎛️ Control variables ({len(value_cols)}): {control_names}")
+        print(f"    Media channels ({len(impression_cols)}): {media_names}")
+        print(f"    Control variables ({len(value_cols)}): {control_names}")
 
         # Get unique regions and weeks
         regions = sorted(df[region_col].unique())
@@ -106,7 +106,7 @@ def load_real_mmm_data(filepath="examples/data/MMM Data.csv"):
         n_regions = len(regions)
         n_weeks = len(weeks)
 
-        print(f"   📊 Data structure: {n_regions} regions × {n_weeks} weeks")
+        print(f"    Data structure: {n_regions} regions × {n_weeks} weeks")
 
         # Create complete grid
         complete_index = pd.MultiIndex.from_product([regions, weeks], names=[region_col, time_col])
@@ -130,7 +130,7 @@ def load_real_mmm_data(filepath="examples/data/MMM Data.csv"):
         df_complete = df_complete.sort_values(['region_idx', 'week_idx'])
 
         # Seasonality features will be added by UnifiedDataPipeline
-        print("   🌊 Seasonality features will be added by UnifiedDataPipeline for proper train/holdout alignment")
+        print("    Seasonality features will be added by UnifiedDataPipeline for proper train/holdout alignment")
         
         # Extract arrays
         X_media_list = []
@@ -148,20 +148,20 @@ def load_real_mmm_data(filepath="examples/data/MMM Data.csv"):
         X_control = np.stack(X_control_list, axis=0)
         y = np.stack(y_list, axis=0)
 
[…]
