### 2025-10-23 9f50cbe
Fix paper.md: Update DOI to concept DOI, replace LightweightMMM with Meridian, remove visualization counts
- Changed DOI from 10.5281/zenodo.17274024 to 10.5281/zenodo.16934842 (concept DOI)
- Replaced LightweightMMM reference with accurate Meridian description
- Removed '14+ interactive dashboards' claim from Summary and Implementation Details
- Removed '(28 tests)' from Reproducibility section
- Updated State of the Field section to accurately describe Meridian's causal inference approach
