+++ JOSS/paper.md
@@ -17,7 +17,7 @@ affiliations:
     name: "Independent Researcher"    
 date: 5 October 2025
 bibliography: paper.bib
-archive_doi: 10.5281/zenodo.17274024
+archive_doi: 10.5281/zenodo.16934842
 ---
 
 # Summary
@@ -26,7 +26,7 @@ Marketing Mix Modeling (MMM) is a statistical technique used to estimate the imp
 
 **DeepCausalMMM** is a Python package that addresses these limitations by combining deep learning, causal inference, and advanced marketing science. The package uses Gated Recurrent Units (GRUs) to automatically learn temporal patterns such as adstock (carryover effects) and lag, while simultaneously learning statistical dependencies and potential causal structures between marketing channels through Directed Acyclic Graph (DAG) learning [@Zheng2018NOTEARS; @Gong2024CausalMMM]. Additionally, it implements Hill equation-based saturation curves to model diminishing returns and optimize budget allocation.
 
-Key innovations include: (1) a data-driven design where hyperparameters and transformations (e.g., adstock decay, saturation curves) are learned or estimated from data with sensible defaults, rather than requiring fixed heuristics or manual specification, (2) multi-region modeling with both shared and region-specific parameters, (3) robust statistical methods including Huber loss and advanced regularization, (4) comprehensive response curve analysis for understanding channel saturation, and (5) an extensive visualization suite with 14+ interactive dashboards for business insights.
+Key innovations include: (1) a data-driven design where hyperparameters and transformations (e.g., adstock decay, saturation curves) are learned or estimated from data with sensible defaults, rather than requiring fixed heuristics or manual specification, (2) multi-region modeling with both shared and region-specific parameters, (3) robust statistical methods including Huber loss and advanced regularization, (4) comprehensive response curve analysis for understanding channel saturation.
 
 # Statement of Need
 
@@ -40,13 +40,13 @@ Several open-source MMM frameworks exist, each with distinct approaches:
 
 **Robyn (Meta)** [@Runge2024RobynPackaging; @RobynGitHub] uses Bayesian hyperparameter optimization with fixed adstock and saturation transformations (Adstock, Hill, Weibull). It provides budget optimization and is widely used in industry but requires manual specification of transformation types and does not model channel interdependencies.
 
-**LightweightMMM (Google)** [@LightweightMMM2022] implements Bayesian MMM using JAX and Numpyro, offering probabilistic inference with flexible priors. It supports adstock effects and budget optimization but does not incorporate causal graph learning or deep learning for temporal dynamics.
+**Meridian (Google)** [@Meridian2024] is Google's open-source Bayesian MMM framework featuring reach and frequency modeling, geo-level analysis, and experimental calibration. It employs causal inference with pre-specified causal graphs and the backdoor criterion. .
 
 **PyMC-Marketing** [@PyMCMarketing2024] provides Bayesian MMM with highly flexible prior specifications and some causal identification capabilities. It excels at uncertainty quantification but requires significant Bayesian modeling expertise and does not use neural networks for temporal modeling.
 
-**CausalMMM** [@Gong2024CausalMMM] introduces neural networks and graph learning to MMM, demonstrating the value of discovering channel interdependencies. However, it does not provide multi-region modeling, comprehensive response curve analysis, or the extensive visualization and analysis tools needed for practical deployment.
+**CausalMMM** [@Gong2024CausalMMM] introduces neural networks and graph learning to MMM, demonstrating the value of discovering channel interdependencies. However, it does not provide multi-region modeling or comprehensive response curve analysis.
 
[…]
