### 2025-10-23 11e2c2d
Revise DOI and enhance paper content
- Updated DOI to concept DOI: 10.5281/zenodo.16934842
- Replaced LightweightMMM with accurate Meridian description
- Added Chan2017 citation to Summary section
- Removed '14+ interactive dashboards' claims
- Removed '(28 tests)' reference
- Capitalized 'kpi' to 'KPI' for consistency
