### 2025-11-16 3c99dbc
feat: Add budget optimization module and clean up documentation
Major Changes:
- Add budget optimization module with BudgetOptimizer class
- Add optimization utility functions for data preparation and reporting
- Add example_budget_optimization.py with complete workflow
- Add API documentation for optimization module
- Remove all emojis from documentation for professional presentation
- Clean up project structure documentation

New Features:
- Budget optimization with multiple methods (SLSQP, trust-constr, differential_evolution, hybrid)
- Channel-specific constraints for business requirements
[…]
