+++ deepcausalmmm/postprocess/optimization.py
@@ -1,0 +1,730 @@
+"""
+Budget Optimization for Marketing Mix Modeling.
+
+This module provides budget optimization capabilities using response curves
+from DeepCausalMMM. It uses constrained optimization to find the optimal
+allocation of marketing budget across channels to maximize predicted response.
+
+The optimizer uses Hill equation saturation curves fitted by ResponseCurveFit
+to predict channel responses at different spend levels and finds the allocation
+that maximizes total response subject to budget and business constraints.
+"""
+
+from typing import Dict, List, Optional, Tuple, Union
+from dataclasses import dataclass
+import logging
+
+import numpy as np
+import pandas as pd
+import scipy.optimize as op
+
+logger = logging.getLogger('deepcausalmmm')
+
+
+@dataclass
+class OptimizationResult:
+    """
+    Result from budget optimization.
+    
+    Attributes
+    ----------
+    success : bool
+        Whether optimization converged successfully
+    allocation : Dict[str, float]
+        Optimal spend allocation by channel
+    predicted_response : float
+        Total predicted response at optimal allocation
+    by_channel : pd.DataFrame
+        Detailed results by channel with spend, response, and ROI
+    message : str
+        Optimization status message
+    method : str
+        Optimization method used
+    
+    Examples
+    --------
+    >>> result = optimizer.optimize()
+    >>> if result.success:
+    ...     print(f"Optimal allocation: {result.allocation}")
+    ...     print(f"Expected response: {result.predicted_response:,.0f}")
+    ...     print(result.by_channel)
+    """
+    success: bool
+    allocation: Dict[str, float]
+    predicted_response: float
+    by_channel: pd.DataFrame
+    message: str = ""
+    method: str = "trust-constr"
+
+
+class BudgetOptimizer:
+    """
+    Optimize marketing budget allocation using response curves.
+    
+    Uses constrained optimization (trust-constr, SLSQP, or differential evolution)
+    with Hill transformation curves from ResponseCurveFit to find optimal spend
+    allocation that maximizes total response subject to business constraints.
+    
+    Parameters
+    ----------
+    budget : float
+        Total budget to allocate across all channels
+    channels : List[str]
+        List of channel names to include in optimization
+    response_curves : Dict[str, Dict]
+        Response curve parameters by channel from ResponseCurveFit.
+        Each channel dict should contain: 'top', 'bottom', 'saturation', 'slope'
+    num_weeks : int, default=52
+        Number of weeks for planning horizon (annual by default)
+    method : str, default='trust-constr'
+        Optimization method: 'trust-constr', 'SLSQP', 'differential_evolution', 'hybrid'
+        
+    Attributes
+    ----------
+    constraints_df : pd.DataFrame or None
+        DataFrame with channel-level constraints (lower, upper bounds)
+        
+    Examples
+    --------
+    >>> # After fitting response curves with ResponseCurveFit
+    >>> curves = {
+    ...     'TV': {'top': 1000000, 'bottom': 0, 'saturation': 50000, 'slope': 1.5},
+    ...     'Search': {'top': 800000, 'bottom': 0, 'saturation': 30000, 'slope': 2.0},
+    ...     'Social': {'top': 600000, 'bottom': 0, 'saturation': 20000, 'slope': 1.8}
+    ... }
+    >>> 
+    >>> optimizer = BudgetOptimizer(
+    ...     budget=1000000,
+    ...     channels=['TV', 'Search', 'Social'],
+    ...     response_curves=curves,
+    ...     num_weeks=52
+    ... )
+    >>> 
+    >>> # Optional: Set channel-specific constraints
+    >>> optimizer.set_constraints({
[…]
