+++ deepcausalmmm/postprocess/optimization_utils.py
@@ -1,0 +1,532 @@
+"""
+Utility functions for budget optimization with DeepCausalMMM.
+
+This module provides helper functions to prepare data from DeepCausalMMM
+model outputs for budget optimization, including data formatting,
+curve parameter extraction, and integration with ResponseCurveFit.
+"""
+
+from typing import Dict, List, Optional, Tuple
+import logging
+
+import pandas as pd
+import numpy as np
+
+from .response_curves import ResponseCurveFit
+from .optimization import BudgetOptimizer, OptimizationResult
+
+logger = logging.getLogger('deepcausalmmm')
+
+
+def prepare_optimization_data(
+    contributions_df: pd.DataFrame,
+    media_data: pd.DataFrame,
+    *,
+    date_col: str = 'week_monday',
+    channel_col: str = 'channel',
+    contribution_col: str = 'predicted',
+    spend_col: str = 'spend',
+    impressions_col: str = 'impressions'
+) -> pd.DataFrame:
+    """
+    Prepare data from DeepCausalMMM outputs for response curve fitting and optimization.
+    
+    This function merges model contribution predictions with media spend/impression
+    data to create the required format for ResponseCurveFit.
+    
+    Parameters
+    ----------
+    contributions_df : pd.DataFrame
+        Model contributions output with columns: date, channel, predicted
+    media_data : pd.DataFrame
+        Media data with columns: date, channel, spend, impressions
+    date_col : str, default='week_monday'
+        Name of the date column
+    channel_col : str, default='channel'
+        Name of the channel column
+    contribution_col : str, default='predicted'
+        Name of the contribution/prediction column
+    spend_col : str, default='spend'
+        Name of the spend column
+    impressions_col : str, default='impressions'
+        Name of the impressions column
+        
+    Returns
+    -------
+    pd.DataFrame
+        Merged data ready for ResponseCurveFit with columns:
+        week_monday, channel, spend, impressions, predicted
+        
+    Examples
+    --------
+    >>> # After training DeepCausalMMM model
+    >>> contributions = model.get_contributions()  # Your model output
+    >>> media_df = pd.read_csv('media_data.csv')
+    >>> 
+    >>> optimization_data = prepare_optimization_data(
+    ...     contributions_df=contributions,
+    ...     media_data=media_df
+    ... )
+    """
+    # Ensure date column is datetime
+    contributions_df = contributions_df.copy()
+    media_data = media_data.copy()
+    
+    contributions_df[date_col] = pd.to_datetime(contributions_df[date_col])
+    media_data[date_col] = pd.to_datetime(media_data[date_col])
+    
+    # Merge contributions with media data
+    merged = pd.merge(
+        contributions_df[[date_col, channel_col, contribution_col]],
+        media_data[[date_col, channel_col, spend_col, impressions_col]],
+        on=[date_col, channel_col],
+        how='inner'
+    )
+    
+    # Rename to standard column names for ResponseCurveFit
+    merged = merged.rename(columns={
+        date_col: 'week_monday',
+        channel_col: 'channel',
+        contribution_col: 'predicted',
+        spend_col: 'spend',
+        impressions_col: 'impressions'
+    })
+    
+    # Remove rows with zero or negative spend/impressions (can't fit curves)
+    merged = merged[
+        (merged['spend'] > 0) & 
+        (merged['impressions'] > 0) & 
+        (merged['predicted'] >= 0)
+    ]
+    
+    logger.info(
+        f"Prepared optimization data: {len(merged)} rows, "
+        f"{merged['channel'].nunique()} channels"
+    )
+    
+    return merged
+
+
+def fit_response_curves_batch(
+    data: pd.DataFrame,
+    channels: Optional[List[str]] = None,
+    *,
[…]

[…]
