+++ test_linear_scaling_100epochs.py
@@ -1,0 +1,268 @@
+#!/usr/bin/env python3
+"""
+Test script for linear scaling (y/y_mean) implementation
+Run with 100 epochs to quickly verify the changes work correctly
+"""
+
+import sys
+import os
+
+# Ensure we're using the local package
+sys.path.insert(0, '/Users/adityapu/Documents/GitHub/deepcausalmmm')
+
+print("="*80)
+print("TESTING LINEAR SCALING IMPLEMENTATION (y/mean)")
+print("="*80)
+print(f"Running quick 100-epoch test to verify architecture changes")
+print("")
+
+# Import after path is set
+from deepcausalmmm.core.config import get_default_config
+from deepcausalmmm.core.trainer import ModelTrainer
+from deepcausalmmm.core.data import UnifiedDataPipeline
+import numpy as np
+import pandas as pd
+
+# Load data
+print("Loading data...")
+df = pd.read_csv('/Users/adityapu/Documents/GitHub/deepcausalmmm/examples/data/MMM Data.csv')
+
+# Extract data
+impression_cols = [col for col in df.columns if 'impressions_' in col]
+if 'target_visits' in df.columns:
+    target_col = 'target_visits'
+    value_cols = [col for col in df.columns if col.startswith('control_')]
+else:
+    target_col = 'value_visits_visits'
+    value_cols = [col for col in df.columns if 'value_' in col and col != 'value_visits_visits']
+
+region_col = 'dmacode'
+time_col = 'weekid'
+
+regions = sorted(df[region_col].unique())
+weeks = sorted(df[time_col].unique())
+n_regions = len(regions)
+n_weeks = len(weeks)
+
+print(f"Data: {n_regions} regions × {n_weeks} weeks")
+print(f"Media channels: {len(impression_cols)}")
+print(f"Control variables: {len(value_cols)}")
+
+# Create complete grid and fill missing values
+complete_index = pd.MultiIndex.from_product([regions, weeks], names=[region_col, time_col])
+complete_df = pd.DataFrame(index=complete_index).reset_index()
+df_complete = complete_df.merge(df, on=[region_col, time_col], how='left')
+
+for col in impression_cols:
+    df_complete[col] = df_complete[col].fillna(0)
+
+for col in value_cols + [target_col]:
+    df_complete[col] = df_complete.groupby(region_col)[col].fillna(method='ffill').fillna(method='bfill')
+    if df_complete[col].isna().any():
+        df_complete[col] = df_complete[col].fillna(df_complete[col].mean())
+
+# Create mappings
+region_map = {region: i for i, region in enumerate(regions)}
+week_map = {week: i for i, week in enumerate(weeks)}
+df_complete['region_idx'] = df_complete[region_col].map(region_map)
+df_complete['week_idx'] = df_complete[time_col].map(week_map)
+df_complete = df_complete.sort_values(['region_idx', 'week_idx'])
+
+# Extract arrays
+X_media_list = []
+X_control_list = []
+y_list = []
+
+for region_idx in range(n_regions):
+    region_data = df_complete[df_complete['region_idx'] == region_idx].sort_values('week_idx')
+    X_media_list.append(region_data[impression_cols].values.astype(np.float32))
+    X_control_list.append(region_data[value_cols].values.astype(np.float32))
+    y_list.append(region_data[target_col].values.astype(np.float32))
+
+X_media = np.stack(X_media_list, axis=0)
+X_control = np.stack(X_control_list, axis=0)
+y = np.stack(y_list, axis=0)
+
[…]
