### 2026-01-25 63d54b3
feat: Replace log-space with linear scaling (y/y_mean) for clean attribution
- Replace log1p transformation with linear scaling by region mean (y/y_mean)
- Update inverse_transform_contributions to use simple multiplication
- Fix seasonality double-counting: store baseline_without_seasonal separately
- Add prediction_scale to inverse transform for correct component scaling
- Pass pipeline to trainer.train() for scaler access
- Update metrics calculation to use scaler.inverse_transform_target()
- Add comprehensive test script (test_linear_scaling_100epochs.py)

Results (100 epochs test):
- ✓ Ad
[…]
