+++ JOSS/paper.md
@@ -24,7 +24,7 @@ archive_doi: 10.5281/zenodo.16934842
 
 Marketing Mix Modeling (MMM) is a statistical technique used to estimate the impact of marketing activities on business outcomes such as sales, revenue, or customer visits. Traditional MMM approaches often rely on linear regression or Bayesian hierarchical models that assume independence between marketing channels and struggle to capture complex temporal dynamics and non-linear saturation effects [@Chan2017; @Hanssens2005; @Ng2021Bayesian].
 
-**DeepCausalMMM** is a Python package that addresses these limitations by combining deep learning, causal inference, and advanced marketing science. The package uses Gated Recurrent Units (GRUs) to automatically learn temporal patterns such as adstock (carryover effects) and lag, while simultaneously learning statistical dependencies and potential causal structures between marketing channels through Directed Acyclic Graph (DAG) learning [@Zheng2018NOTEARS; @Gong2024CausalMMM]. Additionally, it implements Hill equation-based saturation curves to model diminishing returns and optimize budget allocation.
+**DeepCausalMMM** is a Python package that addresses these limitations by combining deep learning, causal inference, and advanced marketing science. The package uses Gated Recurrent Units (GRUs) to automatically learn temporal patterns such as adstock (carryover effects) and lag, while simultaneously learning statistical dependencies between marketing channels through Directed Acyclic Graph (DAG) structure using an upper triangular mask to enforce acyclicity [@Zheng2018NOTEARS; @Gong2024CausalMMM]. Additionally, it implements Hill equation-based saturation curves to model diminishing returns and optimize budget allocation.
 
 Key features include: (1) a data-driven design where hyperparameters and transformations (e.g., adstock decay, saturation curves) are learned or estimated from data with sensible defaults, rather than requiring fixed heuristics or manual specification, (2) multi-region modeling with both shared and region-specific parameters, (3) robust statistical methods including Huber loss and advanced regularization, (4) comprehensive response curve analysis for understanding channel saturation.
 
@@ -46,27 +46,25 @@ Several open-source MMM frameworks exist, each with distinct approaches:
 
 **CausalMMM** [@Gong2024CausalMMM] introduces neural networks and graph learning to MMM, demonstrating the value of discovering channel interdependencies. However, it does not provide multi-region modeling or comprehensive response curve analysis.
 
-**DeepCausalMMM** advances the field by integrating: (1) GRU-based temporal modeling, (2) DAG-based structure learning [@Zheng2018NOTEARS], (3) Hill equation response curves, (4) multi-region modeling, (5) robust statistical methods, (6) production-ready architecture.
+**DeepCausalMMM** advances the field by integrating: (1) GRU-based temporal modeling, (2) DAG-based structure learning using upper triangular constraints [@Zheng2018NOTEARS], (3) Hill equation response curves, (4) multi-region modeling, (5) robust statistical methods, (6) production-ready architecture.
 
-# Functionality
+# Software Design
 
-## Core Architecture
+DeepCausalMMM's architecture reflects several key design decisions driven by the unique challenges of marketing mix modeling:
 
-**Temporal Modeling**: A GRU network automatically learns adstock effects, lag patterns, and time-varying coefficients.
+**Neural Architecture Choice**: We selected Gated Recurrent Units (GRUs) over alternatives (LSTMs, Transformers, simple RNNs) after evaluating trade-offs: GRUs provide sufficient temporal modeling capacity for weekly marketing data while requiring fewer parameters than LSTMs, reducing overfitting risk in typical MMM datasets (50-200 weeks). Transformers were rejected due to their quadratic memory complexity and tendency to overfit on small temporal sequences.
 
[…]
