+++ examples/dashboard_with_counterfactual.py
@@ -618,7 +618,6 @@ def create_beautiful_dashboard():
             holdout_tensors['X_media'], holdout_tensors['X_control'],
             holdout_tensors['R'], holdout_tensors['y'],
             # y_full_for_baseline=y_full_scaled,  # REMOVED: Causes data leakage
-            pipeline=pipeline,
             verbose=True
         )
     else:
@@ -626,7 +625,6 @@ def create_beautiful_dashboard():
             train_tensors['X_media'], train_tensors['X_control'],
             train_tensors['R'], train_tensors['y'],
             # y_full_for_baseline=y_full_scaled,  # REMOVED: Causes data leakage
-            pipeline=pipeline,
             verbose=True
         )
         
@@ -713,6 +711,153 @@ def create_beautiful_dashboard():
     print(f"    Holdout R²: {results['final_holdout_r2']:.3f}")
     print(f"    Performance Gap: {results['final_train_r2'] - results['final_holdout_r2']:.3f} ({((results['final_train_r2'] - results['final_holdout_r2'])/results['final_train_r2']*100):.1f}%)")
     print(f"    All data processed with consistent transformations")
+    
+    # COUNTERFACTUAL ATTRIBUTION TESTING
+    print(f"\n" + "=" * 80)
+    print(f"COUNTERFACTUAL ATTRIBUTION TEST")
+    print(f"=" * 80)
+    print(f"Testing what the trained model predicts with different components zeroed out...")
+    
+    def predict_counterfactual(X_media_input, X_control_input, model, pipeline, n_weeks_orig):
+        """Helper function to predict with counterfactual inputs"""
+        # Add seasonality features
+        X_control_with_seasonality = pipeline._add_seasonality_features(
+            torch.tensor(X_control_input), start_week=0
+        )
+        
+        # Transform
+        X_media_scaled, X_control_scaled, _ = pipeline.scaler.transform(
+            X_media_input, X_control_with_seasonality.numpy(), np.zeros_like(X_media_input[:, :, 0])
+        )
+        
+        # Add padding
+        X_media_padded, X_control_padded, _ = pipeline._add_padding(
+            X_media_scaled, X_control_scaled, torch.zeros(X_media_scaled.shape[0], X_media_scaled.shape[1])
+        )
+        
+        # Create region tensor
+        n_regions_cf = X_media_padded.shape[0]
+        R = torch.arange(n_regions_cf, dtype=torch.long)
+        
+        # Predict
+        model.eval()
+        with torch.no_grad():
+            y_pred_log, _, _, _ = model(X_media_padded, X_control_padded, R)
+        
+        # Remove padding and inverse transform
+        burn_in = config['burn_in_weeks']
+        y_pred_log_eval = y_pred_log[:, burn_in:burn_in+n_weeks_orig]
+        y_pred_orig = torch.expm1(torch.clamp(y_pred_log_eval, max=20.0))
+        
+        return y_pred_orig.sum().item()
+    
+    # Test 1: Full prediction (already have this)
+    total_actual = y.sum()
+    print(f"\nTest 1: Full Model (All Components):")
+    print(f"   Total Actual Visits:            {total_actual:,.0f}")
+    
+    # Use the postprocess predictions for full model
+    total_full_pred = postprocess_results['predictions'].sum().item()
+    print(f"   Full Model Prediction:          {total_full_pred:,.0f} visits")
+    print(f"   Prediction Accuracy:            {(total_full_pred / total_actual * 100):.1f}%")
+    
+    # Test 2: Zero Media
+    print(f"\nTest 2: ZERO Media (Baseline + Controls Only):")
+    X_media_zero = np.zeros_like(X_media)
[…]
