+++ test_release_v1_0_19.py
@@ -1,0 +1,420 @@
+#!/usr/bin/env python
+"""
+Comprehensive Release Test for DeepCausalMMM v1.0.19
+====================================================
+
+Tests all critical functionality before production release.
+"""
+
+import sys
+import traceback
+from pathlib import Path
+
+def test_imports():
+    """Test all core imports."""
+    print("\n" + "="*80)
+    print("TEST 1: Package Imports")
+    print("="*80)
+    
+    try:
+        import deepcausalmmm
+        from deepcausalmmm import DeepCausalMMM, get_device
+        from deepcausalmmm.core import get_default_config, update_config
+        from deepcausalmmm.core.trainer import ModelTrainer
+        from deepcausalmmm.core.data import UnifiedDataPipeline
+        from deepcausalmmm.core.scaling import SimpleGlobalScaler
+        from deepcausalmmm.postprocess import (
+            ComprehensiveAnalyzer,
+            ResponseCurveFit,
+            BudgetOptimizer,
+            optimize_budget_from_curves
+        )
+        from deepcausalmmm.utils.data_generator import ConfigurableDataGenerator
+        
+        print(f"   Package version: {deepcausalmmm.__version__}")
+        print(f"   Device: {get_device()}")
+        print("   SUCCESS: All imports successful")
+        return True
+    except Exception as e:
+        print(f"   FAILED: {e}")
+        traceback.print_exc()
+        return False
+
+def test_version():
+    """Test version number is 1.0.19."""
+    print("\n" + "="*80)
+    print("TEST 2: Version Number")
+    print("="*80)
+    
+    try:
+        import deepcausalmmm
+        version = deepcausalmmm.__version__
+        
+        if version == "1.0.19":
+            print(f"   Version: {version}")
+            print("   SUCCESS: Version is 1.0.19")
+            return True
+        else:
+            print(f"   FAILED: Version is {version}, expected 1.0.19")
+            return False
+    except Exception as e:
+        print(f"   FAILED: {e}")
+        traceback.print_exc()
+        return False
+
+def test_data_generation():
+    """Test synthetic data generation."""
+    print("\n" + "="*80)
+    print("TEST 3: Data Generation")
+    print("="*80)
+    
+    try:
+        import numpy as np
+        from deepcausalmmm.core import get_default_config
+        from deepcausalmmm.utils.data_generator import ConfigurableDataGenerator
+        
+        config = get_default_config()
+        config['random_seed'] = 42
+        
+        generator = ConfigurableDataGenerator(config)
+        X_media, X_control, y = generator.generate_mmm_dataset(
+            n_regions=10,
+            n_weeks=52,
+            n_media_channels=5,
+            n_control_channels=3
+        )
+        
+        assert X_media.shape == (10, 52, 5), f"Wrong X_media shape: {X_media.shape}"
+        assert X_control.shape == (10, 52, 3), f"Wrong X_control shape: {X_control.shape}"
+        assert y.shape == (10, 52), f"Wrong y shape: {y.shape}"
+        
+        print(f"   X_media: {X_media.shape}")
+        print(f"   X_control: {X_control.shape}")
+        print(f"   y: {y.shape}")
+        print("   SUCCESS: Data generation works")
+        return True
+    except Exception as e:
+        print(f"   FAILED: {e}")
+        traceback.print_exc()
+        return False
+
+def test_pipeline():
+    """Test UnifiedDataPipeline."""
+    print("\n" + "="*80)
+    print("TEST 4: Data Pipeline")
+    print("="*80)
+    
+    try:
+        import numpy as np
+        from deepcausalmmm.core import get_default_config
+        from deepcausalmmm.core.data import UnifiedDataPipeline
+        from deepcausalmmm.utils.data_generator import ConfigurableDataGenerator
+        
[…]
