+++ examples/dashboard_linear_scaling.py
@@ -1,418 +1,0 @@
-#!/usr/bin/env python3
-"""
-Dashboard with Linear Scaling (y/y_mean per DMA)
-================================================
-Testing linear scaling approach for proper attribution:
-- Scale: y_scaled = y / y_mean_per_dma
-- Inverse: y_orig = y_scaled * y_mean_per_dma
-- Attribution: Uses PROPORTIONAL ALLOCATION (same as scaling.py):
-  * Calculate component ratios in scaled space
-  * Apply ratios to original-space prediction
-  * Guarantees components sum to 100%!
-"""
-
-import numpy as np
-import torch
-import torch.nn as nn
-from deepcausalmmm.core.config import get_default_config
-from deepcausalmmm.core.unified_model import DeepCausalMMM
-
-def load_real_mmm_data(filepath="examples/data/MMM Data.csv"):
-    """Load and process the real MMM Data.csv"""
-    import pandas as pd
-    
-    print(f" Loading Real MMM Data from: {filepath}")
-    
-    df = pd.read_csv(filepath)
-    print(f"    Loaded data shape: {df.shape}")
-    
-    # Identify columns
-    impression_cols = [col for col in df.columns if 'impressions_' in col]
-    
-    if 'target_visits' in df.columns:
-        target_col = 'target_visits'
-        value_cols = [col for col in df.columns if col.startswith('control_')]
-    else:
-        target_col = 'value_visits_visits'
-        value_cols = [col for col in df.columns if 'value_' in col and col != 'value_visits_visits']
-    
-    region_col = 'dmacode'
-    time_col = 'weekid'
-    
-    # Clean names
-    media_names = [col.replace('impressions_', '').split('_delayed')[0].split('_exponential')[0].split('_geometric')[0].replace('_', ' ') for col in impression_cols]
-    control_names = [col.replace('control_', 'Control ').replace('value_', '').replace('econmetricsmsa_', '').replace('mortgagemetrics_', '').replace('moodys_', '').replace('_sm', '').replace('_', ' ').title() for col in value_cols]
-    
-    print(f"    Media channels ({len(impression_cols)}): {media_names}")
-    print(f"    Control variables ({len(value_cols)}): {control_names}")
-    
-    # Get regions and weeks
-    regions = sorted(df[region_col].unique())
-    weeks = sorted(df[time_col].unique())
-    n_regions = len(regions)
-    n_weeks = len(weeks)
-    
-    print(f"    Data structure: {n_regions} regions × {n_weeks} weeks")
-    
-    # Create complete grid and fill missing
-    complete_index = pd.MultiIndex.from_product([regions, weeks], names=[region_col, time_col])
-    complete_df = pd.DataFrame(index=complete_index).reset_index()
-    df_complete = complete_df.merge(df, on=[region_col, time_col], how='left')
-    
-    for col in impression_cols:
-        df_complete[col] = df_complete[col].fillna(0)
-    
-    for col in value_cols + [target_col]:
-        df_complete[col] = df_complete.groupby(region_col)[col].fillna(method='ffill').fillna(method='bfill')
-        if df_complete[col].isna().any():
-            df_complete[col] = df_complete[col].fillna(df_complete[col].mean())
-    
-    # Create mappings
-    region_map = {region: i for i, region in enumerate(regions)}
-    week_map = {week: i for i, week in enumerate(weeks)}
[…]

[…]
