### 2026-01-25 0da36fa
Release v1.0.19: Linear Scaling Architecture & JOSS Paper Fixes
Major Features:
- Replaced log-space modeling with linear scaling (y/y_mean) for additive attribution
- Implemented attribution prior regularization with dynamic loss scaling
- Added data-driven Hill parameter initialization from SOV percentiles
- Enabled perfect component additivity (sum to 100%)

Bug Fixes:
- Fixed data leakage in scaling.py (y_mean calculation)
- Fixed seasonal initialization broadcasting
- Removed unnecessary y_full_for_baseline parameter from trainer.py
- Fixed integration tests (pipeline parameter, data tran
[…]
