+++ JOSS/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "DeepCausalMMM: A Deep Learning Framework for Marketing Mix Modeling with Causal Inference"
+title: "DeepCausalMMM: A Deep Learning Framework for Marketing Mix Modeling with Causal Structure Learning"
 tags:
   - Python
   - marketing mix modeling
@@ -26,13 +26,13 @@ Marketing Mix Modeling (MMM) is a statistical technique used to estimate the imp
 
 **DeepCausalMMM** is a Python package that addresses these limitations by combining deep learning, causal inference, and advanced marketing science. The package uses Gated Recurrent Units (GRUs) to automatically learn temporal patterns such as adstock (carryover effects) and lag, while simultaneously learning statistical dependencies between marketing channels through Directed Acyclic Graph (DAG) structure using an upper triangular mask to enforce acyclicity [@Zheng2018NOTEARS; @Gong2024CausalMMM]. Additionally, it implements Hill equation-based saturation curves to model diminishing returns and optimize budget allocation.
 
-Key features include: (1) a data-driven design where hyperparameters and transformations (e.g., adstock decay, saturation curves) are learned or estimated from data with sensible defaults, rather than requiring fixed heuristics or manual specification, (2) multi-region modeling with both shared and region-specific parameters, (3) robust statistical methods including Huber loss and advanced regularization, (4) comprehensive response curve analysis for understanding channel saturation.
+Key features include: (1) a data-driven design where hyperparameters and transformations (e.g., adstock decay, saturation curves, Hill parameters) are learned or estimated from data with sensible defaults, (2) linear scaling architecture enabling additive attribution where components sum to 100%, (3) configurable attribution prior regularization with dynamic loss scaling for business-aligned allocations, (4) multi-region modeling with both shared and region-specific parameters, (5) robust statistical methods including Huber loss and advanced regularization, (6) comprehensive response curve analysis for understanding channel saturation.
 
 # Statement of Need
 
 Marketing organizations invest billions annually in advertising across channels (TV, digital, social, search), yet measuring ROI remains challenging due to: (1) temporal complexity with delayed and persistent effects [@Hanssens2005], (2) channel interdependencies [@Gong2024CausalMMM], (3) non-linear saturation with diminishing returns [@Li2024Survey], (4) regional heterogeneity, and (5) multicollinearity between campaigns.
 
-**DeepCausalMMM** addresses these challenges by combining GRU-based temporal modeling, DAG-based structure learning, Hill equation response curves, multi-region modeling, production-ready performance (91.8% holdout R², 3.0% train-test gap), and data-driven hyperparameter learning for generalizability.
+**DeepCausalMMM** addresses these challenges by combining GRU-based temporal modeling, DAG-based structure learning, Hill equation response curves, multi-region modeling, production-ready performance (83.9% holdout R², 10.8% train-test gap), realistic attribution through configurable prior regularization, and data-driven hyperparameter learning for generalizability.
 
 # State of the Field
 
@@ -46,25 +46,31 @@ Several open-source MMM frameworks exist, each with distinct approaches:
 
 **CausalMMM** [@Gong2024CausalMMM] introduces neural networks and graph learning to MMM, demonstrating the value of discovering channel interdependencies. However, it does not provide multi-region modeling or comprehensive response curve analysis.
 
-**DeepCausalMMM** advances the field by integrating: (1) GRU-based temporal modeling, (2) DAG-based structure learning using upper triangular constraints [@Zheng2018NOTEARS], (3) Hill equation response curves, (4) multi-region modeling, (5) robust statistical methods, (6) production-ready architecture.
[…]
