### 2026-01-25 a9a3a32
Remove obsolete wait_for_dashboards.sh debugging script
