### 2026-01-25 623b8cf
Fix performance metrics in JOSS paper
Updated to reflect actual final 1500-epoch training results:
- Training R²: 0.947 → 0.956 (correct)
- Holdout R²: 0.839 (unchanged)
- Gap: 10.8% → 12.2% (correct calculation)
- Removed RMSE lines (outdated from earlier run)

Word count: 1,564 (reduced from 1,582, still within 1,750 limit)
