### 2026-01-25 82cf69d
Clarify linear scaling description in JOSS paper
Updated to specify:
- y/y_mean scaling is applied per region
- Applied to the dependent variable
- More precise technical description

Word count: 1,572 (still within 1,750 limit)
