### 2026-01-25 8da53c3
Fix README.md: Correct DOI and update API examples
Fixes:
1. DOI: Updated from 10.5281/zenodo.17274024 to 10.5281/zenodo.16934842 (matches paper.md)
2. Basic Usage example: Updated to use correct v1.0.19 API
   - temporal_split() for train/holdout split
   - fit_and_transform_training() for data processing
   - trainer.train() with explicit tensors and pipeline parameter
   - Removed deprecated fit_transform() and old train() signature

Example now matches paper.md and quickstart.ipynb
