+++ README.md
@@ -5,7 +5,7 @@
 [![Documentation](https://readthedocs.org/projects/deepcausalmmm/badge/?version=latest)](https://deepcausalmmm.readthedocs.io/en/latest/?badge=latest)
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/adityapt/deepcausalmmm/blob/main/examples/quickstart.ipynb)
 [![PyPI version](https://badge.fury.io/py/deepcausalmmm.svg)](https://badge.fury.io/py/deepcausalmmm)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17274024.svg)](https://doi.org/10.5281/zenodo.17274024)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.16934842.svg)](https://doi.org/10.5281/zenodo.16934842)
 [![MMM](https://img.shields.io/badge/Marketing%20Mix-Modeling-brightgreen)](https://en.wikipedia.org/wiki/Marketing_mix_modeling)
 [![Deep Learning](https://img.shields.io/badge/Deep-Learning-blue)](https://pytorch.org)
 [![Causal DAG](https://img.shields.io/badge/Causal-DAG-purple)](https://en.wikipedia.org/wiki/Directed_acyclic_graph)
@@ -67,32 +67,57 @@ pip install torch pandas numpy plotly networkx statsmodels scikit-learn tqdm
 ### Basic Usage
 
 ```python
-import pandas as pd
-from deepcausalmmm import DeepCausalMMM, get_device
+import numpy as np
+from deepcausalmmm import get_device
 from deepcausalmmm.core import get_default_config
 from deepcausalmmm.core.trainer import ModelTrainer
 from deepcausalmmm.core.data import UnifiedDataPipeline
 
-# Load your data
-data = pd.read_csv('your_mmm_data.csv')
+# Prepare your data (example with numpy arrays)
+# X_media: [n_regions, n_weeks, n_media_channels]
+# X_control: [n_regions, n_weeks, n_control_variables]
+# y: [n_regions, n_weeks]
 
-# Get optimized configuration
+# Get configuration
 config = get_default_config()
+config['n_epochs'] = 200  # Adjust as needed
 
 # Check device availability
 device = get_device()
 print(f"Using device: {device}")
 
-# Process data with unified pipeline
+# Initialize pipeline
 pipeline = UnifiedDataPipeline(config)
-processed_data = pipeline.fit_transform(data)
 
-# Train with ModelTrainer (recommended approach)
+# Split data temporally (train/holdout)
+train_data, holdout_data = pipeline.temporal_split(X_media, X_control, y)
+
+# Process training data
+train_tensors = pipeline.fit_and_transform_training(train_data)
+holdout_tensors = pipeline.transform_holdout(holdout_data)
+
+# Create trainer and model
 trainer = ModelTrainer(config)
-model, results = trainer.train(processed_data)
+model = trainer.create_model(
+    n_media=train_tensors['X_media'].shape[2],
+    n_control=train_tensors['X_control'].shape[2],
+    n_regions=train_tensors['X_media'].shape[0]
+)
+trainer.create_optimizer_and_scheduler()
 
-# Generate comprehensive dashboard
-python dashboard_rmse_optimized.py  # Run the main dashboard script
+# Train model
+results = trainer.train(
+    train_tensors['X_media'], train_tensors['X_control'],
+    train_tensors['R'], train_tensors['y'],
+    holdout_tensors['X_media'], holdout_tensors['X_control'],
+    holdout_tensors['R'], holdout_tensors['y'],
[…]
