### 2026-01-25 4b6fb73
Update package structure in README.md
Added missing files/folders:
- test_release_v1_0_19.py (comprehensive release test)
- RELEASE_NOTES_1.0.19.md (latest release notes)
- examples/data/ directory with MMM Data.csv

Structure now accurately reflects v1.0.19 package contents
