### 2026-01-25 d2dfb8f
Fix performance benchmarks in README.md
Updated to match JOSS paper and final 1500-epoch training results:
- Training R²: 0.947 → 0.956 (correct)
- Holdout R²: 0.839 (unchanged)
- Performance Gap: 12.2% (correct calculation)
- Removed outdated RMSE values

Now consistent with paper.md metrics
