### 2026-01-26 2e1532f
Update CITATION.cff to v1.0.19
Updated version from 1.0.18 to 1.0.19 in both locations:
- Main version field
- Preferred citation version field

DOI remains correct (10.5281/zenodo.16934842)
