### 2026-01-26 7fa7a8c
Remove test_release_v1_0_19.py (user request)
