### 2026-01-26 05f5822
Update README.md: Remove test_release_v1_0_19.py from structure
