### 2026-01-26 927a664
Add roadmap section: NOTEARS implementation in v1.0.20 (Q2 2025)
Added Roadmap section announcing:
- Full NOTEARS DAG learning implementation planned for v1.0.20
- Enhanced causal discovery beyond upper triangular constraints
- Target release: Q2 2025
