### 2026-01-26 93fbf7c
Fix: Add complete data generation example in README.md
- Added ConfigurableDataGenerator import and usage
- Included concrete synthetic data creation (50 regions, 104 weeks, 13 media, 3 controls)
- Fixed incomplete example that would have thrown NameError
- Now provides fully executable code that users can run immediately
