### 2026-01-26 4affbff
Fix: Correct dashboard command path and clarify examples require repo clone
- Changed 'python dashboard_rmse_optimized.py' to 'python examples/dashboard_rmse_optimized.py'
- Added clarification that examples/ folder is not included in pip install
- Provided full workflow: clone repo -> install package -> run examples
- Added other example commands for completeness
