### 2026-01-26 66af4b7
Fix: Add missing data definitions in ResponseCurveFit and optimize_budget examples
- ResponseCurveFit: Added channel_data creation with week, impressions, contributions
- optimize_budget_from_curves: Added fitted_curves_df DataFrame with sample parameters
- Both examples now have complete, executable code without NameError
- Added comments to guide users on replacing with actual model data
