### 2026-01-26 1a7277f
Fix: Correct all code examples in README.md to match actual API
Tested all examples and fixed:
1. ConfigurableDataGenerator: Use generate_mmm_dataset() method with correct parameters
2. ResponseCurveFit: Use correct DataFrame columns (week_monday, spend, impressions, predicted)
3. ResponseCurveFit: Use fit() method instead of non-existent fit_curve() and calculate_r2_and_plot()
4. All 4 code examples now tested and verified to work correctly

Changes:
- Basic Usage: Fixed ConfigurableDataGenerator API (n_regions -> generate_mmm_dataset())
- Response Curves: Fixed DataFrame columns and method c
[…]
