### 2026-01-26 6ca24b8
Add version compatibility note for README examples
- Added clear note that examples are for v1.0.19+ (feature branch)
- Updated GitHub installation to specify feature/remove-log-scaling branch
- Addresses confusion for users on v1.0.18 from PyPI
- Recommends cloning repo or waiting for v1.0.19 release
