### 2026-01-26 275dc29
Add critical API compatibility warning for v1.0.18 vs v1.0.19
- Added prominent warning section explaining version incompatibility
- Listed specific API changes between versions
- Provided 3 clear options for users on v1.0.18
- Highlighted that PyPI examples will NOT work for v1.0.18
- Addresses user confusion about TypeError with 'pipeline' parameter
