### 2026-01-26 46ed4bb
Remove emojis from README.md version compatibility section
- Removed warning emoji from section header
- Removed check marks and X marks from API changes list
- Maintains clear, professional formatting
