### 2026-01-26 212c0af
Rename FINAL_RELEASE_SUMMARY to RELEASE_NOTES and remove all emojis
- Replaced FINAL_RELEASE_SUMMARY_v1.0.19.md with more comprehensive version
- Removed all emojis from release notes (user preference)
- Converted checkmarks to plain bullets where appropriate
- Changed 'Built with ❤️' to 'Built for' (emoji-free)
