### 2026-01-26 108c62d
Clean up README: Remove trend/data leakage mentions, simplify Hill initialization quote
- Removed 'Trend: 0% (frozen as requested)' line
- Removed 'No data leakage' mention (already implied by proper implementation)
- Simplified quote to focus on data-driven Hill initialization feature
- Changed 'Zero hardcoding approach...' to 'Data-driven Hill initialization enables channel-specific saturation curves'
