### 2026-01-26 60d9fef
Polish JOSS paper for final review: Soften causal claims, improve language
Changes for JOSS reviewer acceptance:
- Tag: 'causal inference' → 'causal structure learning' (aligns with actual claims)
- Figure 1 caption: 'potential causal relationships' → 'statistical dependencies consistent with plausible causal pathways'
- 'production-ready performance' → 'production-oriented performance characteristics including strong temporal generalization'
- '(indicating strong generalization)' → '(indicating reasonable generalization under a strict temporal holdout)'
- Multi-region modeling: Added 'This de
[…]
