### 2026-01-26 9601e95
Final JOSS polish: Add complementary positioning, versioning policy, soften performance claims
Recommended edits for reviewer acceptance:
- State of Field: Added 'DeepCausalMMM is complementary to Bayesian MMM frameworks...'
- Implementation: Added semantic versioning and backward compatibility statement
- Performance: Changed 'has demonstrated strong performance' to 'In an applied real-world marketing analytics use case, DeepCausalMMM achieved'
- Performance: Added 'These results illustrate practical viability rather than serving as a controlled benchmark comparison'

All changes align with J
[…]
