### 2026-01-26 2243d18
Remove log-space metrics from README feature list
Log-space metrics no longer relevant with v1.0.19 linear scaling architecture
