### 2026-01-26 7f718d5
Make reproducibility explicit and actionable for JOSS reviewers
Critical reproducibility improvements:
1. Note on Benchmarks: Specify exact dataset path (examples/data/MMM Data.csv) and workflow
2. Research Impact: Tightened 'reproducible benchmarks' claim to 'fully reproducible benchmark workflow'
3. Reproducibility section: Complete rewrite with actionable details
   - Explicit dataset location
   - Exact command to run: python examples/dashboard_rmse_optimized.py
   - Output location specified
   - Answers reviewer question: 'What do I run, and where is the data?'

Word count: 1,709 words (
[…]
