### 2026-01-26 d2f3731
Final editorial polish: Provenance, PII disclaimer, author voice consistency
Editorial improvements for JOSS submission:
1. Added provenance statement to COI section (independent work, personal time)
2. Added PII disclaimer to benchmark data description
3. Made author voice consistent throughout (We → The author)

These changes prevent avoidable reviewer back-and-forth on editorial details.

Word count: 1,754 words (at JOSS limit)
