### 2026-01-26 9fe4638
Trim word count to 1,736 (under 1,750 JOSS limit)
Removed 18 words from non-essential phrases:
- 'solely for the purpose of reproducing the results reported here' → 'for reproducibility' (8 words)
- 'under strict temporal holdout' removed (4 words)
- 'using an anonymized real-world dataset and executable scripts included in the repository' → 'with included dataset and executable scripts' (6 words)

Word count: 1,736 / 1,750 (98.4% of limit, 14 words headroom)
