### 2026-01-26 a49aeb1
Update all metrics to match dashboard output for reproducibility
Critical reproducibility fix:
- Training R²: 0.956 → 0.950
- Holdout R²: 0.839 → 0.842
- Performance Gap: 12.2% (0.117) → 10.8% (0.108)

Updated files:
- JOSS/paper.md (Performance & Research Impact sections)
- README.md (Performance Benchmarks)
- RELEASE_NOTES_1.0.19.md (Production Performance)
- CHANGELOG.md (Performance section)

Reason: These metrics match what reviewers will see when running:
python examples/dashboard_rmse_optimized.py

Word count unchanged: 1,736 words (within JOSS 1,750 limit)
