### 2026-01-26 3fcf1dc
Fix critical mathematical error and clarify PyPI distribution status
Three critical fixes:
1. Gap calculation: '10.8% relative to training R²' → '10.8 percentage points'
   - Previous: Incorrectly implied 10.8% relative (actual: 11.4%)
   - Corrected: Clear absolute gap of 10.8 percentage points (0.950 - 0.842)
   - More intuitive for practitioners

2. PyPI distribution clarity: 'Distributed via PyPI' → 'Available via PyPI (v1.0.19 release concurrent with this publication)'
   - Clarifies version alignment with reviewed code

3. Removed redundant '0.108' mention (already stated as percentage p
[…]
