### 2026-01-26 ed2ec6b
Add comprehensive end-to-end MMM tutorial
New Tutorial: docs/source/tutorials/basic_mmm_workflow.rst
- Complete workflow from data generation to budget optimization
- Self-contained with synthetic data (no external files needed)
- Covers all key features: training, attribution, response curves, optimization
- 8 step-by-step sections with executable code examples
- Troubleshooting tips and best practices
- ~400 lines of comprehensive documentation

Updated: docs/source/tutorials/index.rst
- Linked to new tutorial
- Removed placeholder 'Coming soon' sections
- Cleaner structure

Ready for JOSS r
[…]
