+++ docs/source/tutorials/basic_mmm_workflow.rst
@@ -1,0 +1,511 @@
+End-to-End MMM Analysis
+========================
+
+This tutorial walks through a complete Marketing Mix Modeling workflow using DeepCausalMMM, from data generation to budget optimization.
+
+Overview
+--------
+
+In this tutorial, you'll learn how to:
+
+1. Generate synthetic MMM data
+2. Configure and train a model
+3. Analyze attribution results
+4. Fit response curves
+5. Optimize budget allocation
+
+This entire workflow uses synthetic data, so you can run it immediately without any external data files.
+
+Step 1: Setup and Data Generation
+----------------------------------
+
+First, let's import necessary libraries and generate synthetic MMM data:
+
+.. code-block:: python
+
+    import numpy as np
+    import pandas as pd
+    from deepcausalmmm import get_device
+    from deepcausalmmm.core import get_default_config
+    from deepcausalmmm.core.trainer import ModelTrainer
+    from deepcausalmmm.core.data import UnifiedDataPipeline
+    from deepcausalmmm.utils.data_generator import ConfigurableDataGenerator
+    
+    # Set random seed for reproducibility
+    np.random.seed(42)
+    
+    # Initialize synthetic data generator
+    generator = ConfigurableDataGenerator(
+        base_sales=50000,
+        n_media_channels=5,
+        n_control_vars=3,
+        trend_strength=0.0,
+        seasonality_strength=0.3
+    )
+    
+    # Generate MMM dataset
+    # Returns: X_media [regions, weeks, channels], X_control, y [regions, weeks]
+    X_media, X_control, y = generator.generate_mmm_dataset(
+        n_regions=10,
+        n_weeks=104
+    )
+    
+    print(f"Generated data shapes:")
+    print(f"  Media: {X_media.shape}")      # (10, 104, 5)
+    print(f"  Control: {X_control.shape}")  # (10, 104, 3)
+    print(f"  Target: {y.shape}")           # (10, 104)
+
+Understanding the Data
+~~~~~~~~~~~~~~~~~~~~~~
+
+Our synthetic dataset represents:
+
+- **10 regions** (e.g., different DMAs or geographic markets)
+- **104 weeks** (~2 years of data)
+- **5 media channels** (e.g., TV, Search, Social, Display, Radio)
+- **3 control variables** (e.g., price, promotion, competitor spend)
+- **Target variable** (e.g., sales, revenue, or conversions)
+
+Step 2: Configure the Model
+----------------------------
+
+DeepCausalMMM is highly configurable. Let's start with the default configuration and customize key parameters:
+
+.. code-block:: python
+
+    # Get default configuration
+    config = get_default_config()
+    
+    # Customize for our use case
+    config['n_epochs'] = 500  # Reduce for faster training (default: 1500)
+    config['learning_rate'] = 0.001
+    
+    # Attribution prior regularization (optional but recommended)
+    config['media_contribution_prior'] = 0.40  # Target 40% media attribution
+    config['attribution_reg_weight'] = 0.5     # Balanced regularization
+    
+    # Check device
+    device = get_device()
+    print(f"Training on: {device}")
+
+Key Configuration Parameters
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- **n_epochs**: Number of training iterations (500-1500)
+- **media_contribution_prior**: Target percentage for media attribution (0.30-0.50)
+- **attribution_reg_weight**: Regularization strength (0.0-1.0)
+- **holdout_ratio**: Validation split percentage (default: 0.12)
+
+Step 3: Prepare Data Pipeline
+------------------------------
+
[…]
