### 2026-01-26 38015bf
Fix reproducibility command in JOSS paper
Critical Fix: Added complete setup instructions for benchmark reproduction

Before (incomplete):
  python examples/dashboard_rmse_optimized.py

After (complete):
  git clone https://github.com/puttaparthy-aditya/deepcausalmmm.git
  cd deepcausalmmm
  pip install -e .
  python examples/dashboard_rmse_optimized.py

This ensures JOSS reviewers can actually reproduce results without errors.
README.md already had complete instructions - now JOSS paper matches.
