### 2026-01-26 de8d928
Refine JOSS paper language and technical descriptions
Three key refinements for academic rigor:

1. Linear Scaling Description (Line 65):
   - Changed from simple 'y/y_mean' explanation
   - Now includes formal mathematical notation (y/ȳ_r)
   - Explicitly connects to econometric decomposition models
   - Clarifies scale invariance and additive properties

2. Performance Claims (Line 164):
   - Softened 'strong out-of-sample predictive accuracy'
   - Changed to 'reasonable out-of-sample predictive accuracy'
   - More measured language appropriate for JOSS standards

3. Software Features (Line 
[…]
