### 2026-01-26 45110f5
Revise README for installation and version updates
Updated installation instructions and version compatibility notes in README. Clarified changes related to version 1.0.19 and improved phrasing for better clarity.
