### 2026-01-26 d7c67fb
Remove 'Unreleased' section from CHANGELOG
Removed the 'Unreleased' section from the CHANGELOG.
