+++ CHANGELOG.md
@@ -5,8 +5,6 @@ All notable changes to DeepCausalMMM will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
 ## [1.0.19] - 2026-01-25
 
 ### Linear Scaling & Attribution Architecture
