### 2026-01-26 a9e2362
Fix ConfigurableDataGenerator API usage in tutorial
Critical Bug Fix: Incorrect API usage in documentation

Problem:
  Tutorial incorrectly passed parameters to ConfigurableDataGenerator.__init__():
    generator = ConfigurableDataGenerator(base_sales=..., n_media_channels=...)

  This caused: TypeError: __init__() got an unexpected keyword argument 'base_sales'

Correct API:
  ConfigurableDataGenerator.__init__() takes only optional 'config' parameter
  All dataset parameters go to generate_mmm_dataset() method

Fixed in docs/source/tutorials/basic_mmm_workflow.rst:
  - Removed incorrect para
[…]
