### 2026-01-26 e162118
Fix critical bugs in end-to-end tutorial
Multiple Critical Fixes:

1. Model Forward Call (Step 5):
   - Fixed: model() returns TUPLE not dict
   - Correct: y_pred, media_coeffs, media_contrib_direct, outputs = model(...)
   - Was causing: TypeError: tuple indices must be integers or slices, not str

2. Correct Dictionary Keys (Step 5):
   - Fixed: outputs['contributions'] not outputs['media_contribution']
   - Fixed: outputs['control_contributions'] not outputs['control_contribution']
   - Fixed: outputs['seasonal_contribution'] (singular)
   - Fixed: Use y_pred from tuple, not outputs['y_pred
[…]
