### 2026-01-26 371addd
Fix all broken examples in quickstart guide
Critical Fixes for 3 Out of 4 Examples:

EXAMPLE 2: Model Inference
- Old: Used InferenceManager (has bugs in v1.0.19)
- New: Direct model inference with torch.no_grad()
- Fixed: Correct model forward() tuple unpacking
- Fixed: Correct dictionary keys (outputs['contributions'])

EXAMPLE 3: Response Curves Analysis
- Old: Wrong parameters (x_col, y_col don't exist)
- Old: Wrong method (fit_curve() doesn't exist)
- New: Correct column names ('week_monday', 'spend', 'impressions', 'predicted')
- New: Correct method: fit() instead of fit_curve()
- Added:
[…]
