### 2026-01-26 16d6b70
Fix Step 6 (Response Curves) and dashboard path
Step 6 (Response Curves) - Complete Rewrite:
- Old API: Used non-existent parameters (x_col, y_col, model_level='national')
- Old API: Called non-existent method fit_curve()
- New API: Correct column names ('week_monday', 'spend', 'impressions', 'predicted')
- New API: Correct method fit() returns DataFrame
- New API: model_level='Overall' (not 'national')
- Added: Better error handling and warning message
- Added: Explanation for why fitting might fail on synthetic data

Dashboard Path Fix (quickstart.rst):
- Old: python dashboard_rmse_optimized
[…]
