### 2026-01-26 25a667f
Fix Step 6 array length mismatch (ValueError)
Critical Fix: Step 6 Response Curves - Array Length Mismatch

Root Cause:
- X_media: original data (104 weeks)
- train_data['X_media']: training split (92 weeks)
- media_contrib: model output with padding (98 weeks = 92 + 6 burn-in)
- Mismatch caused: ValueError: All arrays must be of the same length

Solution:
- Use train_data['X_media'] instead of X_media
- Remove burn-in padding from media_contrib before aggregating
- Now: channel_spend and channel_contrib both have 92 weeks ✅

Code Changes:

Verified: No more ValueError, arrays match correctly.
[…]
