### 2026-01-26 bdb82cc
Add clarification that curve fitting failures are OK in Step 6
User-Requested Clarification:

Added multiple notes that curve fitting failures are expected and acceptable:

1. Upfront note before code block:
   'Curve fitting may not succeed with limited training epochs or synthetic data.
    This is expected and OK - the code handles it gracefully.'

2. In error messages:
   Changed 'failed' to 'skipped' and added '(OK)' suffix

3. In warning message:
   Changed 'Warning' to 'Note'
   Emphasized 'this is OK and expected'
   Added context about quick tutorial runs vs production

This prevents 
[…]
