### 2026-01-26 e98a22d
Remove all emojis from core package
Removed 4 emojis from unified_model.py:

Line 428: 📊 -> (removed)
  OLD: logger.info(f'\n📊 Initializing Hill g parameters from SOV data:')
  NEW: logger.info(f'\nInitializing Hill g parameters from SOV data:')

Line 462: ✅ -> (removed)
  OLD: logger.info(f'\n✅ Hill g initialization complete')
  NEW: logger.info(f'\nHill g initialization complete')

Line 937: ⚠️ -> (removed)
  OLD: print(f'⚠️  WARNING: Components don't sum to raw_prediction!')
  NEW: print(f'WARNING: Components don't sum to raw_prediction!')

Line 939: ⚠️ -> (removed)
  OLD: print(f'⚠️  WARNI
[…]
