### 2026-01-27 5d6e811
Refine language and clarity in JOSS paper
Updated terminology and clarified descriptions in the paper to enhance readability and accuracy.
