### 2026-01-27 d343346
Update MANIFEST.in to include CSV examples
