### 2026-01-27 e5157be
fix: Remove all log scaling references and fix example_response_curves.py
- Add CSV files to MANIFEST.in for JOSS reproducibility
- Fix example_response_curves.py: replace torch.expm1() with correct inverse_transform_target()
- Fix example_response_curves.py: correct API call to inverse_transform_contributions()
- Remove vestigial log_transform parameter from scaling.py
- Update all comments from 'log space' to 'scaled space (y/y_mean)'
- Rename 30+ variables from _log to _scaled for clarity
- All 28 tests passing (24 unit + 4 integration)
