### 2026-01-27 4678643
Update RELEASE_NOTES_1.0.19 to remove testing details
Removed testing section and related test file reference from release notes.
