### 2026-01-27 c5048fd
docs: Trim JOSS paper to 1745 words (under 1750 limit)
- Condensed Summary section
- Tightened Key features list
- Removed redundant phrases
- All technical content preserved
