+++ JOSS/paper.md
@@ -22,11 +22,11 @@ archive_doi: 10.5281/zenodo.16934842
 
 # Summary
 
-Marketing Mix Modeling (MMM) is a statistical technique used to estimate the impact of marketing activities on business outcomes such as sales, revenue, or customer visits. Traditional MMM approaches often rely on linear regression or Bayesian hierarchical models that assume independence between marketing channels and struggle to capture complex temporal dynamics and non-linear saturation effects [@Chan2017; @Hanssens2005; @Ng2021Bayesian].
+Marketing Mix Modeling (MMM) estimates the impact of marketing activities on business outcomes such as sales or revenue. Traditional MMM approaches rely on linear regression or Bayesian hierarchical models that assume channel independence and struggle to capture temporal dynamics and non-linear saturation [@Chan2017; @Hanssens2005; @Ng2021Bayesian].
 
-**DeepCausalMMM** is a Python package that addresses these limitations by combining deep learning, causal inference, and advanced marketing science. The package uses Gated Recurrent Units (GRUs) to automatically learn temporal patterns such as adstock (carryover effects) and lag, while simultaneously learning statistical dependencies between marketing channels through Directed Acyclic Graph (DAG) structure using an upper triangular mask to enforce acyclicity [@Zheng2018NOTEARS; @Gong2024CausalMMM]. Additionally, it implements Hill equation-based saturation curves to model diminishing returns and optimize budget allocation.
+**DeepCausalMMM** addresses these limitations by combining deep learning, causal inference, and marketing science. It uses Gated Recurrent Units (GRUs) to learn temporal patterns (adstock, lag) while learning statistical dependencies between channels through Directed Acyclic Graph (DAG) structure with upper triangular constraints [@Zheng2018NOTEARS; @Gong2024CausalMMM]. It implements Hill equation saturation curves for diminishing returns and budget optimization.
 
-Key features include: (1) a data-driven design where hyperparameters and transformations (e.g., adstock decay, saturation curves, Hill parameters) are learned or estimated from data with sensible defaults, (2) linear scaling architecture enabling additive attribution where components sum to 100%, (3) configurable attribution prior regularization with dynamic loss scaling for business-aligned allocations, (4) multi-region modeling with both shared and region-specific parameters, (5) robust statistical methods including Huber loss and advanced regularization, (6) comprehensive response curve analysis for understanding channel saturation.
+Key features: (1) data-driven hyperparameters learned from data with defaults, (2) linear scaling enabling additive attribution (components sum to 100%), (3) configurable attribution priors with dynamic loss scaling, (4) multi-region modeling with shared and region-specific parameters, (5) robust methods including Huber loss, (6) response curve analysis.
 
 # Statement of Need
 
