### 2026-01-27 1d34adc
Revise key features and AI usage disclosure in paper.md
Updated key features and AI usage disclosure sections for clarity and accuracy.
