### 2026-01-27 #9228 @yewentao256
Hi  @Naeemkh  @cjerzak  @salbalkus   @ankurankan  would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "DeepCausalMMM: A Deep Learning Framework for Marketing Mix Modeling with Causal Inference".

Thank you so much!
Wentao
