### 2026-01-29 #9914 @editorialbot
[REVIEW]: DeepCausalMMM: A Deep Learning Framework for Marketing Mix Modeling with Causal Inference
**Submitting author:** @adityapt (<a href="http://orcid.org/0009-0008-9495-3932">Aditya Puttaparthi Tirumala</a>)
**Repository:** https://github.com/adityapt/deepcausalmmm/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.20
