### 2026-02-01 9f223d7
Revise RELEASE_NOTES_1.0.19 for clarity and updates
Updated performance section and fixed formatting issues in release notes.
