### 2026-02-01 e79571b
Refactor README for clarity and consistency
Updated sections in README for clarity and consistency, including renaming headers and refining descriptions of features and benchmarks.
