### 2026-03-14 #9914 @HaoyuanHe0606
Hi @adityapt, thanks for your submission. The package is clearly well thought out and the documentation, testing, and CI setup are all solid. I wanted to flag a few areas from the review checklist that I think need attention before this can move forward.

The biggest gap right now is community engagement. Every single commit in the repo (all 207 of them) comes from you, and there's no evidence of external users filing issues, asking questions, or contributing code. The one merged PR is from your own branch. I get that this is a solo project and that's fine, but JOSS does expect to see some sign that people outside the project have interacted with it. Even a handful of issues from users trying the package, or discussions about feature requests, would go a long way. I'd suggest posting about the package in relevant communities (PyData, marketing analytics forums, etc.) and making it easy for people to file issues by adding templates. If colleagues or beta testers gave you feedback that shaped the design, consider documenting that somewhere visible.

The repo spans about 6.5 months from first to last commit, which technically meets the minimum for development history. But the commit pattern is quite bursty. There are days with 30+ commits and then long gaps. Reviewers may wonder whether this reflects sustained, iterative development or rapid code generation. If you did prototyping or design work before the public repo existed, it would help to mention that in your response to reviewers or in the paper itself. Your AI Usage Disclosure is refreshingly honest, and the Software Design section shows real thought went into the architecture, so I think you're in decent shape here. Just be prepared for questions.

The current Research Impact Statement mostly restates what the package can do rather than providing evidence that it has had impact or is positioned for near term significance. JOSS wants to see things like: has anyone cited the arXiv preprint? Has the package been used in coursework or industry analysis? Are there download numbers from PyPI worth mentioning? Most importantly, a comparative benchmark against an existing tool like Robyn or PyMC-Marketing on the same dataset would make a much stronger case. Even a simple table showing R², RMSE, and runtime side by side would demonstrate that DeepCausalMMM offers something the existing ecosystem doesn't.

On a related note, the paper reports Training R² of 0.950 and Holdout R² of 0.842, but only on the single included dataset with no comparison to other frameworks. These numbers are hard to interpret in isolation. Running one of the established MMM packages on your included MMM Data.csv and showing a side by side comparison would be the most straightforward fix. It doesn't need to be exhaustive, just enough to give reviewers context for what the numbers mean relative to the field.

One small thing: Granger1969 and LightweightMMM2022 are in paper.bib but don't appear to be cited anywhere in paper.md. JOSS requires that all bib entries are actually referenced. Either work them into the text where they're relevant or remove them.

I want to be clear that a lot of this submission is strong. The license, PyPI packaging, Sphinx docs on ReadTheDocs, the test suite running across Python 3.9 through 3.13, the contributing guidelines, the reproducibility setup with an included dataset and executable script, and the Software Design section in the paper are all exactly what JOSS looks for. The State of the Field section does a good job differentiating from existing tools, and the AI disclosure is transparent and appropriate. The core technical contribution here is real. Addressing the community engagement and benchmarking gaps would put this in a much stronger position. Happy to discuss any of this further.
