### 2026-03-15 #9914 @adityapt
Hey @HaoyuanHe0606 -

Thank you for the thorough and constructive review.

**Next steps I will take:**

**Bibliography:** I will ensure every entry in paper.bib is cited in the manuscript. I will either add citations for Granger1969 and LightweightMMM2022 in the relevant sections or remove them so that all references in the bibliography are cited. I will most likely remove them from the references.

**Benchmarking:** I will add a comparative benchmark on the same dataset (MMM Data.csv) and temporal train/holdout split using an existing MMM framework (using PyMC-Marketing and Robyn if applicable). I will add a comparison table (R², RMSE, runtime) and a short narrative in the Performance section so the reported metrics can be interpreted relative to the field. I've been using part of this in my professional work at Zillow, but I shouldn't be sharing those numbers externally.

**Research Impact Statement:** I will revise the Research Impact Statement to emphasize evidence of impact and near-term significance. The revision will incorporate the comparative benchmark, the gap the package fills in the open-source MMM ecosystem, and who benefits (practitioners and researchers), rather than restating the package’s features.

**Community engagement**

If you notice in the package, I have added GitHub issue templates (Bug report, Feature request, Question) to make it easier for users to file issues. I am uncertain how much external engagement JOSS expects for a solo project where I cannot generate issues or contributions myself. I can promote the package in relevant communities (e.g. PyData, marketing analytics forums) to invite feedback; beyond that, engagement depends on others choosing to try the package and report back. I would welcome guidance from you or the editor on whether the steps above are sufficient, or if the submission should wait until there is more visible community activity.

**Development history**

I will add a brief note in the ReadMe.MD that a substantial part of the work was designed in my local system and prototyped before the public repository existed on GitHub, so the commit history reflects integration and documentation of that prior work rather than 6.5 months of linear development.

Thanks for the feedback and am happy to make any further adjustments you or the editor suggest.
