### 2026-03-15 #9914 @Naeemkh
Dear @yewentao256,

Thank you again for inviting me to review the manuscript. I sincerely apologize for not being able to complete the review as I initially indicated. Due to unexpected time constraints on my end, I was unable to give the manuscript the careful attention it deserves.

I appreciate your understanding and your decision to invite additional reviewers to keep the process moving forward. At this point, I believe it would be best for me to step aside from the review process so that the manuscript can proceed without further delay.

I apologize again for the inconvenience, and I wish you and the authors the best with the review process.

Best regards,
Naeem Khoshnevis
