### 2026-03-15 #9914 @yewentao256
@adityapt Thanks! Your TODOs LGTM

> A history of contributions and engagement from individuals beyond the original team, across organisations, is especially welcome, though not essential.

According to https://joss.theoj.org/about, it is not a must have so don't need to worry too much.

> I can promote the package in relevant communities (e.g. PyData, marketing analytics forums) to invite feedback;

This would certainly help!
