### 2026-03-16 cda2f71
Update paper.bib
Remove Granger1969 and LightweightMMM2022 references
