+++ JOSS/paper.bib
@@ -16,18 +16,6 @@
   url          = {https://papers.nips.cc/paper/8157-dags-with-no-tears-continuous-optimization-for-structure-learning}
 }
 
-@article{Granger1969,
-  author       = {C. W. J. Granger},
-  title        = {Investigating Causal Relations by Econometric Models and Cross-Spectral Methods},
-  journal      = {Econometrica},
-  year         = {1969},
-  volume       = {37},
-  number       = {3},
-  pages        = {424--438},
-  doi          = {10.2307/1912791},
-  url          = {https://www.jstor.org/stable/1912791}
-}
-
 @book{Hanssens2005,
   author       = {Dominique M. Hanssens and Leonard J. Parsons and Randall L. Schultz},
   title        = {Market Response Models: Econometric and Time Series Analysis},
@@ -67,15 +55,6 @@
   version      = {1.2.1},
   year         = {2025},
   note         = {accessed October 2025}
-}
-
-@misc{LightweightMMM2022,
-  author       = {Google LightweightMMM contributors},
-  title        = {LightweightMMM: Marketing Mix Modeling (MMM)},
-  year         = {2022},
-  howpublished = {GitHub repository},
-  note         = {accessed October 2025},
-  url          = {https://github.com/google/lightweight_mmm}
 }
 
 @misc{PyMCMarketing2024,
