### 2026-03-23 d6c080e
Docs: benchmark-only examples; remove standalone PyMC/Robyn/Meridian notebooks
- Update README/CONTRIBUTING for quickstart + three-way benchmark
- Adjust .gitignore notebook exceptions
- Refresh mmm_three_way_benchmark.ipynb; drop redundant example notebooks

Made-with: Cursor
