### 2026-03-23 #9914 @tomzhengy
Ref: adityapt/deepcausalmmm#12

  Ref: adityapt/deepcausalmmm#11
