### 2026-03-27 bd0a434
fix: restore benchmark loader on pandas 3
