### 2026-03-27 3f4fc56
fix: align inference outputs with model contract
