### 2026-03-28 #9914 @tomzhengy
@adityapt Thank you for the submission! Here are my suggestions after an initial look

While setting up and testing, I ran into two main issues:

1.  While on my local device (MacOS 26.1), the paper’s reproduction command, `python examples/dashboard_rmse_optimized.py`, currently fails in a clean environment under `pandas 3.0.1` before training begins because the data-loading path uses `groupby(...).fillna(...)`.

1.1 I was able to resolve this setup problem on my machine in my PR https://github.com/adityapt/deepcausalmmm/pull/12

2. While testing the inference path directly with a minimal model and synthetic inputs, I also found a public API mismatch in `InferenceManager.predict()`. In my clean local check, `return_contributions=True` fails with `ValueError: too many values to unpack (expected 3)`, and `return_contributions=False` also fails. I found that this was due to the full model return tuple being treated as a tensor.

2.1 I also made a PR for this https://github.com/adityapt/deepcausalmmm/pull/11

Some other changes I suggest making for a more complete submission:

3. I found this while reading through the paper: it mentions that the package “maintains backward compatibility guarantees,”. However when I checked the README and changelog explicitly describe breaking changes in v1.0.19 and note that v1.0.18 and below use a different API. This may be confusing to readers and I recommend revising that sentence so it matches the actual release history.

4. The documentation looks great but it currently builds with multiple warnings on my device, including references to missing API/example pages. This is a second priority fix but I think cleaning those up would increase user ease of use.

5. The Sphinx docs reference several pages that don't exist: api/index.rst includes cli and visualization, and examples/index.rst includes retail_mmm and multi_region. The docs also omit scipy from docs/requirements.txt even though response_curves.py and optimization.py import it. I suggest cleaning these up.

6. The installation docs (docs/source/installation.rst) advertise Python 3.8+ and optional extras [dev], [visualization], and [docs], but pyproject.toml requires Python >=3.9 and only defines a [test] extra. Running pip install .[dev] in a clean environment does nothing at the moment because the extra packages don't exist.

Once these are addressed, I think the submission will be in much stronger shape. Great work!
