### 2026-03-28 0ea5a6a
JOSS: benchmark table, impact statement, README dev history
- paper.md: comparative Table 1 (same CSV/split), corrected holdout weeks,
  Performance interpretation, trimmed Research Impact with near-term significance,
  Reproducibility points to mmm_three_way_benchmark.ipynb
- README.md: Development history note (pre-public prototyping, bursty commits)

Made-with: Cursor
