+++ JOSS/paper.md
@@ -149,34 +149,30 @@ print(f"Holdout RMSE original scale: {results['final_holdout_rmse']:.0f}")
 
 # Performance
 
-**Note on Benchmarks:** The performance metrics reported below were generated using the end-to-end workflow in `examples/dashboard_rmse_optimized.py` and an anonymized dataset included in the repository at `examples/data/MMM Data.csv`. This ensures reviewers can reproduce the reported numbers and figures using the public code and data artifacts provided. The dataset contains no personally identifiable information (PII) and is distributed for reproducibility.
+Results use `examples/data/MMM Data.csv` (190 DMAs × 109 weeks, 13 channels, 7 controls; no PII) with `holdout_ratio = 0.12` in `examples/pymc_aligned_dcm_config.json`—about **96** train and **13** holdout weeks of observed time (burn-in padding may apply; see pipeline logs).
 
-In an applied real-world marketing analytics use case, DeepCausalMMM achieved the following results on anonymized data containing 190 geographic regions (DMAs), 109 weeks of observations, 13 marketing channels, and 7 control variables. The model uses a temporal train-holdout split with 101 training weeks (92.7%) and the most recent 8 weeks (7.3%) reserved for out-of-sample validation:
+**Table 1** comes from `examples/mmm_three_way_benchmark.ipynb`: same CSV and split, **sklearn** R²/RMSE (original scale, pooled), and execution time. **PyMC-Marketing** [@PyMCMarketing2024], **Meridian** [@Meridian2024], and **national weekly Ridge** on Robyn-style inputs [@Runge2024RobynPackaging] (not Meta’s full Robyn unless `robynpy` is enabled). Bayesian runs use modest MCMC budgets in the notebook.
 
-- **Training R²**: 0.950, **Holdout R²**: 0.842
-- **Train–holdout gap**: 10.8 percentage points (indicating reasonable generalization)
+| Method | Scope | Train R² | Holdout R² | Holdout RMSE | Execution time (s) |
+|--------|-------|----------|------------|--------------|---------------------|
+| National weekly Ridge (Robyn-style inputs) | National, weekly | 0.856 | −12.43 | 1.7 × 10⁷ | — |
+| DeepCausalMMM (dashboard training path) | Panel, geo×week | 0.949 | 0.843 | 5.3 × 10⁵ | 489 |
+| PyMC-Marketing MMM | Panel, geo×week | 0.994 | 0.903 | 4.8 × 10⁵ | 5995 |
+| Meridian | Panel, geo×week | 0.997 | −10.05 | 5.1 × 10⁶ | 479 |
 
-**Attribution Quality**:
-- Configurable attribution priors enable business-aligned allocations through regularization (e.g., media contribution target: 40%)
-- Dynamic loss scaling ensures regularization has meaningful impact during training
-
-These results illustrate practical viability rather than serving as a controlled benchmark comparison. They demonstrate the model's ability to capture complex marketing dynamics while maintaining reasonable out-of-sample predictive accuracy and realistic attribution through configurable prior-based regularization.
-
-**Key Technical Innovations**: (1) Linear scaling (y/y_mean) for dependent variable, (2) Configurable attribution priors with dynamic loss scaling to prevent unrealistic allocations, (3) Data-driven Hill parameter initialization from channel-specific SOV percentiles, (4) Seasonality based regularization.
[…]
