### 2026-03-28 eccd37c
docs: JOSS paper benchmark + README dev history; changelog
- JOSS/paper.md: Table 1 vs PyMC-Marketing/Meridian/Ridge, split fix, impact, reproducibility
- README.md: Development history section
- CHANGELOG.md: [Unreleased] documentation entries

Made-with: Cursor
