### 2026-03-28 #9914 @adityapt
Hey @HaoyuanHe0606

Thank you again for the detailed review. Based on your review, below are the changes I pushed.

**Comparative benchmark (same dataset and split):** I added an explicit comparison on the repository’s examples/data/MMM Data.csv, using the same temporal holdout as examples/pymc_aligned_dcm_config.json (holdout_ratio = 0.12, with the corrected description of train/holdout weeks in the manuscript). The paper.md now includes a table (Table 1) with sklearn train/holdout R², RMSE (original KPI scale, pooled), and execution time for each of those, alongside PyMC-Marketing, Meridian, a national weekly Ridge baseline on Robyn-style inputs, and DeepCausalMMM. The full workflow is in examples/mmm_three_way_benchmark.ipynb, and # Reproducibility points to that notebook for regenerating the table. The Performance text interprets the table carefully (e.g. national vs panel data scope, modest MCMC settings for the Bayesian runs, and what the Ridge row does and does not represent).

**Research Impact Statement:** I rewrote this section so it no longer reads as a second feature list. It now emphasizes the niche the package fills (PyTorch-oriented, multi-region panel MMM with holdouts and the same public comparison), near-term significance via distribution and tooling (PyPI, Zenodo DOI, Read the Docs, CI across Python 3.9–3.13), limits (early citations and course adoption; private industry results not reported in the paper), and who benefits (practitioners vs researchers).

**Development history / commit pattern:** I added a short Development history subsection to README.md explaining that substantial design and prototyping preceded the public GitHub repository and that the visible history reflects integration, documentation, tests, and packaging rather than a day-by-day log of initial exploration, including why commit activity can appear bursty.

**Bibliography:** I verified that every entry in JOSS/paper.bib is cited in paper.md (or removed unused entries), consistent with JOSS requirements, and removed the ones that are uncited.

**Community engagement:** I understand the concern about external interaction on a solo-maintained project. Beyond making reproduction and filing issues easier (e.g. clear paths in the README and paper), I am continuing to welcome feedback through GitHub and am engaging relevant communities where appropriate; I know visible third-party issues and contributions will take time. Looks like the package has been mentioned in some of the LinkedIn posts by MMM practitioners, but it's too early for me to get full community engagement.

LMK if this addresses your questions in detail. I will check and fix any minor bugs if any.
