### 2026-04-04 0d09a12
docs(test): changelog + groupby(sort=False); tidy dashboard EOF
- Document pandas loader fix and new integration test in CHANGELOG [Unreleased]
- Use groupby(..., sort=False) and inline comment for per-region ffill/bfill
- Remove trailing whitespace on final print; ensure newline at EOF

Made-with: Cursor
