+++ examples/dashboard_rmse_optimized.py
@@ -118,7 +118,8 @@ def load_real_mmm_data(filepath="examples/data/MMM Data.csv"):
             df_complete[col] = df_complete[col].fillna(0)  # No impressions = 0
 
         for col in value_cols + [target_col]:
-            df_complete[col] = df_complete.groupby(region_col)[col].transform(
+            # pandas 2.2+ removed SeriesGroupBy.fillna(method=...); use per-region ffill/bfill
+            df_complete[col] = df_complete.groupby(region_col, sort=False)[col].transform(
                 lambda series: series.ffill().bfill()
             )
             if df_complete[col].isna().any():
@@ -2498,4 +2499,5 @@ if __name__ == "__main__":
     print(f"\n Total plots created: {len(results['plots_created'])}")
     print(f"\n To view the dashboard:")
     print(f"   open {dashboard_dir}/master_dashboard.html")
-    print("="*80 + "\n") 
+    print("="*80 + "\n")
+
