+++ CHANGELOG.md
@@ -11,10 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **`examples/dashboard_rmse_optimized.py`**: Region-wise missing-value fill uses `groupby(...).transform(lambda series: series.ffill().bfill())` instead of deprecated / removed grouped `fillna(method='ffill'|'bfill')`, compatible with pandas 2.2+ and 3.x (including environments where `SeriesGroupBy.fillna` is unavailable).
 - **`DeepCausalMMM.forward()` contract vs callers**: **`InferenceManager.predict()`**, **`UnifiedDataPipeline.predict_and_postprocess`**, and **`train_model`** paths now unpack **`(predictions, media_coeffs, media_contributions, outputs)`** and take **control contributions** from **`outputs['control_contributions']`**, matching the current model implementation (fixes wrong media/control tensors when `n_media ≠ n_control`, broken 3-value unpack, and **`return_contributions=False`** returning a tuple instead of predictions).
 - **`JOSS/paper.bib`**: Wrapped corporate author names in double braces (`{{...}}`) so BibTeX stops reordering them into initials-first form (`Meridian2024`, `PyMCMarketing2024`, `RobynGitHub`); also protected product names (`{Meridian}`, `{PyMC-Marketing}`, `{Robyn}`) in titles. Addresses JOSS review feedback on malformed citations.
+- **`postprocess/analysis.py`**: **`ModelAnalyzer.analyze_predictions`** now calls **`InferenceManager.predict`** with the correct signature (NumPy inputs, optional **`return_media_coefficients=True`** for plots); **`_generate_plots`** uses **`media_contributions`**, **`burn_in_weeks`** from the underlying model, and no longer mis-passes region tensors as keyword arguments.
+- **`examples/example_response_curves.py`**: Forward unpack matches **`(predictions, media_coeffs, media_contributions, outputs)`** with **control** from **`outputs['control_contributions']`**; seasonal slice uses **`outputs['seasonal_contribution']`** (singular key).
+- **`examples/dashboard_rmse_optimized.py`**: Holdout **`model(...)`** unpack uses accurate names / discards for unused middle returns (behavior unchanged; clarity only).
 
 ### Added
 - **`tests/integration/test_dashboard_rmse_optimized.py`**: Regression test that loads `load_real_mmm_data()` on `examples/data/MMM Data.csv` so the dashboard data path stays covered in CI.
-- **`tests/unit/test_inference.py`**: Regression coverage for **`InferenceManager.predict()`** with **`return_contributions`** true/false; integration/unit forward-pass tests assert coefficient vs contribution shapes and **`outputs`** consistency.
+- **`tests/unit/test_inference.py`**: Regression coverage for **`InferenceManager.predict()`** with **`return_contributions`** true/false; integration/unit forward-pass tests assert coefficient vs contribution shapes and **`outputs`** consistency; optional **`return_media_coefficients`** coverage.
+- **`InferenceManager.predict(..., return_media_coefficients=...)`**: Optional time-varying media coefficients in the result dict (single forward pass) for analyzers and tooling.
 
 ### Documentation
 - **`installation.rst`** / **`index.rst`**: Aligned install guidance with **`pyproject.toml`** and README (Python **3.9+**, full runtime dependency list including **scipy**, **networkx**, **tqdm**, note that **numpy** is capped below **2.0** in metadata); **PyPI** first, then GitHub; removed nonexistent **`[dev]`** / **`[visualization]`** / **`[docs]`** extras; documented real **`[test]`** extra and doc build requirements file.
[…]
