### 2026-04-04 802b9e2
fix: align analyzer, examples, and docs with forward() contract
- InferenceManager.predict: single forward unpack; optional return_media_coefficients
- ModelAnalyzer.analyze_predictions: correct predict() API; plots use media tensors + model burn-in
- example_response_curves: control from outputs; seasonal_contribution key
- dashboard: accurate holdout model() unpack names (unused slots)
- DeepCausalMMM.forward docstring/examples match actual returns and output keys
- tests: return_media_coefficients; CHANGELOG [Unreleased]
