+++ tests/unit/test_inference.py
@@ -63,3 +63,26 @@ def test_inference_manager_predict_without_contributions_returns_predictions_onl
 
     assert list(results.keys()) == ['predictions']
     assert np.allclose(results['predictions'], predictions.cpu().numpy())
+
+
+def test_inference_manager_predict_can_return_media_coefficients():
+    model, X_media, X_control, regions = _build_model_and_inputs()
+    with pytest.warns(
+        UserWarning,
+        match="Neither pipeline nor scaler provided",
+    ):
+        manager = InferenceManager(model)
+
+    with torch.no_grad():
+        _, media_coeffs, _, _ = model(X_media, X_control, regions)
+
+    results = manager.predict(
+        X_media.numpy(),
+        X_control.numpy(),
+        return_contributions=False,
+        remove_padding=False,
+        return_media_coefficients=True,
+    )
+
+    assert set(results.keys()) == {'predictions', 'media_coefficients'}
+    assert np.allclose(results['media_coefficients'], media_coeffs.cpu().numpy())
