### 2026-04-04 4d0333f
fix: exclude burn-in from final trainer metrics and align dashboard
- ModelTrainer: trim burn-in in scaled space before inverse transform and R²/RMSE
- Use same trim for holdout scaled-space metrics; add _trim_burn_in_time helper
- Dashboard: holdout scatter matches post-burn-in evaluation; MAE uses holdout_mae_orig
- CHANGELOG [Unreleased]
