### 2026-04-04 738fba2
docs(JOSS): align paper versioning with v1.0.19 breaking changes
- Replace blanket backward-compatibility wording with SemVer + changelog/README pointers
- Note v1.0.19 vs pre-1.0.19 API/scaling differences
- CHANGELOG [Unreleased] documentation entry
