### 2026-04-04 06e3e40
docs: align install guidance, Sphinx API/examples, and README deps
- installation.rst and index.rst: PyPI first, Python 3.9+, real [test] extra, full runtime deps
- Add api/cli.rst and api/visualization.rst; example stubs retail_mmm and multi_region
- docs/requirements.txt: scipy for autodoc
- README/CONTRIBUTING: pyproject-aligned requirements and versioned pip deps line
- JOSS paper: trim optional versioning sentence; CHANGELOG [Unreleased] notes

Made-with: Cursor
