+++ CHANGELOG.md
@@ -16,7 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **`tests/unit/test_inference.py`**: Regression coverage for **`InferenceManager.predict()`** with **`return_contributions`** true/false; integration/unit forward-pass tests assert coefficient vs contribution shapes and **`outputs`** consistency.
 
 ### Documentation
-- **JOSS (`paper.md`)**: **Software Design → Implementation Details**: versioning bullet now states **semantic versioning** and **documented breaking changes** (notably **v1.0.19** vs **v1.0.18 and earlier**), with pointers to README/CHANGELOG—replacing a blanket “backward compatibility guarantees” phrase that conflicted with published release notes.
+- **`installation.rst`** / **`index.rst`**: Aligned install guidance with **`pyproject.toml`** and README (Python **3.9+**, full runtime dependency list including **scipy**, **networkx**, **tqdm**, note that **numpy** is capped below **2.0** in metadata); **PyPI** first, then GitHub; removed nonexistent **`[dev]`** / **`[visualization]`** / **`[docs]`** extras; documented real **`[test]`** extra and doc build requirements file.
+- **`contributing.rst`**: Replaced **`pip install -e .[dev]`** with **`pip install -e .`** and **`pip install -e .[test]`** to match optional dependencies.
+- **`README.md`** / **`CONTRIBUTING.md`**: Development requirements aligned with **`pyproject.toml`**; **“Dependencies only”** pip snippet uses the same **version specifiers** as **`[project] dependencies`**; **`CONTRIBUTING.md`** no longer references nonexistent **`[dev]`** extra; **Benchmarks** temporal split matches default **12%** holdout (~96 / ~13 weeks on 109 observed weeks).
+- **Sphinx**: Added missing **`docs/source/api/cli.rst`** and **`visualization.rst`** (``deepcausalmmm.cli``, ``deepcausalmmm.core.visualization``); added **`docs/source/examples/retail_mmm.rst`** and **`multi_region.rst`** so API and Examples toctrees resolve; **`docs/requirements.txt`** includes **scipy** for autodoc imports used by optimization/response-curve modules.
+- **JOSS (`paper.md`)**: **Software Design → Implementation Details**: versioning bullet states **semantic versioning** and **documented breaking changes** (notably **v1.0.19** vs **v1.0.18 and earlier**), with pointers to README/CHANGELOG—replacing a blanket “backward compatibility guarantees” phrase that conflicted with published release notes; optional minor/patch milestone wording was removed for brevity.
 - **JOSS (`paper.md`)**: Comparative **Table 1** on `examples/data/MMM Data.csv` (same split as `pymc_aligned_dcm_config.json`) versus PyMC-Marketing, Meridian, and a national weekly Ridge baseline (Robyn-style inputs; not Meta’s full Robyn unless `robynpy` is used); corrected train/holdout week description (~96 / ~13 observed weeks at 12% holdout); **Research Impact Statement** reframed (niche, reproducible comparison, near-term significance, honest limits on early uptake); **Reproducibility** references `examples/mmm_three_way_benchmark.ipynb` for Table 1.
 - **README.md**: **Development history** note—substantial design/prototyping predates the public GitHub history; bursty commits reflect integration, docs, tests, and packaging.
 
