+++ docs/source/examples/retail_mmm.rst
@@ -1,0 +1,4 @@
+Retail MMM analysis
+===================
+
+A full narrative retail MMM walkthrough is not published here yet. For working code paths, see :doc:`../quickstart` and the ``examples/`` directory in the repository (for example ``examples/dashboard_rmse_optimized.py`` and ``examples/mmm_three_way_benchmark.ipynb``).
