+++ docs/source/api/visualization.rst
@@ -1,0 +1,7 @@
+Visualization
+=============
+
+.. automodule:: deepcausalmmm.core.visualization
+   :members:
+   :undoc-members:
+   :show-inheritance:
