### 2026-04-04 79f2012
docs(readme): expand project structure for tests and Sphinx pages
- tests/unit: test_inference.py; integration: test_dashboard_rmse_optimized.py
- docs: api cli.rst, visualization.rst; examples retail_mmm.rst, multi_region.rst

Made-with: Cursor
