### 2026-04-05 bd44847
Fix formatting of versioning section in paper.md
