### 2026-04-08 #9914 @tomzhengy
> Hey [@tomzhengy](https://github.com/tomzhengy), thanks for the review, below are the updates based on your feedback:
>
> 1. **python examples/dashboard_rmse_optimized.py**: Fixed the data loading path that used grouped fillna (removed in recent pandas). This matches the fix you proposed in PR which is merged to main.
> 2. **InferenceManager.predict() / return_contributions:** Fixed the public API mismatch. The predictions now unpack the model’s full return tuples, and control contributions come from outputs['control_contributions'] as expected.  This aligns with PR which is merged to main.
> 3. **Paper.md fixes on backward compatibility:** Revised the JOSS paper so it no longer claims backward-compatibility guarantees. It now states versioning, mentions the documented breaking changes, and explicitly references v1.0.19 vs older versions, with pointers to README and CHANGELOG for migration context.
> 4. **Sphinx, missing pages and scipy dependencies:** Added the missing API pages (cli, visualization) and coming soon statement to retail_mmm, multi_region so that people know, and added scipy to docs/requirements.txt so autodoc can import modules that depend on SciPy.
> 5. **Installation docs vs pyproject.toml:** Updated docs, source, installation.rst (and the landing index.rst) so they match pyproject.toml. README and CONTRIBUTING are also aligned the same way.
>
> Also, a note on doc builds in GitHub and local. A clean sphinx build may still emit warnings (largely from documenting the same symbols in both the curated API pages and the recursive autosummary tree). HTML builds succeed; there can be an option to trim warning volume but this is common.
>
> LMK if you see any other issues on your end, and I will be happy to make changes accordingly.

This looks good to me! Updated my checklist and will continue looking at updates.
