### 2026-04-13 #9914 @yewentao256
Hi @adityapt , several citation issues:
- `Team, G. M. M. M. M. (2025)`
- `contributors, P.-M. (2024)`
- `contributors), M. (Robyn. (2024). Robyn: Media mix modeling by meta`

Please take a look and fix them, others LGTM

After the fix, please take a look at "Additional Author Tasks After Review is Complete" mentioned above, thanks!
