### 2026-04-14 #9914 @adityapt
Hey @yewentao256 - I checked the official Google Meridian github, this is how they wanted me to cite it:

https://github.com/google/meridian/blob/main/CITATION.cff

And, that's how I cited it in the bib file (when I accessed it in 2025):

@software{Meridian2024,
  author       = {Google Meridian Marketing Mix Modeling Team},
  title        = {Meridian: Marketing Mix Modeling},
  url          = {https://github.com/google/meridian},
  version      = {1.2.1},
  year         = {2025},
  note         = {accessed October 2025}
}

I also checked pymc-marketing, no citation available: https://github.com/pymc-labs/pymc-marketing/tree/main

So, internet recommended me to cite it like this in my bib file.

@misc{PyMCMarketing2024,
  author       = {PyMC-Marketing contributors},
  title        = {PyMC-Marketing: Open Source Marketing Analytics},
  year         = {2024},
  howpublished = {Project website},
  note         = {accessed October 2025},
  url          = {https://www.pymc-marketing.io}
}

Same with facebook Robyn: https://github.com/facebookexperimental/Robyn

@misc{RobynGitHub,
  author       = {Meta (Robyn contributors)},
  title        = {Robyn: Media Mix Modeling by Meta},
  year         = {2024},
  howpublished = {GitHub repository},
  note         = {accessed October 2025},
  url          = {https://github.com/facebookexperimental/Robyn}
}

These are not academic journals, but real softwares on GitHub. Not sure how else to cite those. Please LMK.
