### 2026-04-14 #9914 @yewentao256
@adityapt Would this work? For example

```bash
@software{Meridian2024,
  author  = {{Google Meridian Marketing Mix Modeling Team}},
  title   = {{Meridian}: Marketing Mix Modeling},
  url     = {https://github.com/google/meridian},
  version = {1.2.1},
  year    = {2025},
  note    = {Accessed October 2025}
}
```

Similar to others, the problem currently is that M. M. M. is not a name of author
