### 2026-04-18 fe6d496
Refactor entries in paper.bib for consistency
Updated author and title fields for consistency and capitalization in multiple entries. Changed 'accessed' to 'Accessed' for uniformity.
