+++ JOSS/paper.bib
@@ -49,29 +49,29 @@
 }
 
 @software{Meridian2024,
-  author       = {Google Meridian Marketing Mix Modeling Team},
-  title        = {Meridian: Marketing Mix Modeling},
+  author       = {{Google Meridian Marketing Mix Modeling Team}},
+  title        = {{Meridian}: Marketing Mix Modeling},
   url          = {https://github.com/google/meridian},
   version      = {1.2.1},
   year         = {2025},
-  note         = {accessed October 2025}
+  note         = {Accessed October 2025}
 }
 
 @misc{PyMCMarketing2024,
-  author       = {PyMC-Marketing contributors},
-  title        = {PyMC-Marketing: Open Source Marketing Analytics},
+  author       = {{PyMC-Marketing contributors}},
+  title        = {{PyMC-Marketing}: Open Source Marketing Analytics},
   year         = {2024},
   howpublished = {Project website},
-  note         = {accessed October 2025},
+  note         = {Accessed October 2025},
   url          = {https://www.pymc-marketing.io}
 }
 
 @misc{RobynGitHub,
-  author       = {Meta (Robyn contributors)},
-  title        = {Robyn: Media Mix Modeling by Meta},
+  author       = {{Robyn contributors (Meta)}},
+  title        = {{Robyn}: Media Mix Modeling by Meta},
   year         = {2024},
   howpublished = {GitHub repository},
-  note         = {accessed October 2025},
+  note         = {Accessed October 2025},
   url          = {https://github.com/facebookexperimental/Robyn}
 }
 
