### 2026-04-18 eef60b8
Release v1.0.20: JOSS review fixes
- paper.bib: brace-protect corporate author names (Meridian, PyMC-Marketing,
  Robyn) so BibTeX stops rendering them as initialized personal names;
  protect product names in titles. Addresses reviewer feedback.
- CITATION.cff: align title with JOSS paper.md; bump version to 1.0.20;
  refresh date-released and preferred-citation metadata.
- pyproject.toml: bump version to 1.0.20.
- CHANGELOG.md: move [Unreleased] bucket into [1.0.20] release section and
  document the bib and CITATION.cff changes.
- RELEASE_NOTES_1.0.20.md: summary of patch-release bug fixes 
[…]
