+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "deepcausalmmm"
-version = "1.0.19"
+version = "1.0.20"
 description = "Causal Deep Learning marketing-mix modeling library with response curves"
 readme = "README.md"
 requires-python = ">=3.9"
