### 2026-04-18 de0089a
docs(readme): switch PyPI version badge to shields.io for faster refresh
Made-with: Cursor
