### 2026-04-18 #9914 @adityapt
Hi @yewentao256 - thanks for the pointers. Posting updates on some of those follow-ups:

1. paper.bib citations (Meridian, PyMC-Marketing, Robyn)

Fixed in JOSS/paper.bib as you recommended, I'll generate the pdf and hopefully it should generate the names correctly.

Also I made sure I double checked the following:

Below are details:

New release version with all the changes v1.0.20 has been pushed to GitHub, PyPi and Zenodo.
Archive DOI (Zenodo): [10.5281/zenodo.16934842](https://doi.org/10.5281/zenodo.16934842) (concept DOI - so, this always resolves to the latest version; v1.0.20 is now the latest archived version and I checked it to be safe)

Please let me know if anything else is needed - happy to fix.
