### 2026-04-19 d0dfa1c
docs(paper.bib): correct Li2024 locator; brace-protect Meta/Robyn's in titles
- Li2024Survey: replace fabricated pages 14999-15023 with the correct
  article number 298 (journal uses article-numbered publishing, not
  paginated issues); switch URL to canonical doi.org resolver.
- RobynGitHub title: brace-protect {Meta} so CSL sentence-case does not
  lowercase the company name.
- Runge2024RobynPackaging title: brace-protect {Robyn's} so CSL does not
  lowercase the product name.

Made-with: Cursor
