+++ JOSS/paper.bib
@@ -43,9 +43,9 @@
   year         = {2024},
   volume       = {57},
   number       = {11},
-  pages        = {14999--15023},
+  pages        = {298},
   doi          = {10.1007/s10462-024-10886-0},
-  url          = {https://link.springer.com/article/10.1007/s10462-024-10886-0}
+  url          = {https://doi.org/10.1007/s10462-024-10886-0}
 }
 
 @software{Meridian2024,
@@ -68,7 +68,7 @@
 
 @misc{RobynGitHub,
   author       = {{Robyn contributors (Meta)}},
-  title        = {{Robyn}: Media Mix Modeling by Meta},
+  title        = {{Robyn}: Media Mix Modeling by {Meta}},
   year         = {2024},
   howpublished = {GitHub repository},
   note         = {Accessed October 2025},
@@ -86,7 +86,7 @@
 }
 
 @article{Runge2024RobynPackaging,
-  title        = {Packaging Up Media Mix Modeling: An Introduction to Robyn's Open‑Source Approach},
+  title        = {Packaging Up Media Mix Modeling: An Introduction to {Robyn's} Open‑Source Approach},
   author       = {Julian Runge and Igor Skokan and Gufeng Zhou and Koen Pauwels},
   journal      = {CoRR},
   volume       = {abs/2403.14674},
