### 2026-04-25 af74021
docs(paper): fix Table 1 overflow and reduce body-text bolding
Reviewer feedback on the JOSS PDF flagged two issues. Both are addressed
in this commit and previewed by a new draft-PDF GitHub Action.

- JOSS/paper.md: shorten Table 1 method/scope cells and column headers
  ("DeepCausalMMM (dashboard training path)" -> "DeepCausalMMM",
  "Panel, geo x week" -> "Panel", "Execution time (s)" -> "Time (s)")
  and move scope/inputs context into a table caption to prevent the
  leftmost column from overflowing the page.
- JOSS/paper.md: drop ~50 emphasis bolds in the Performance, Research
  Impact, R
[…]
