### 2026-04-25 ac6e376
Refine text for clarity and consistency in paper.md
Corrected punctuation and phrasing for clarity in the analysis of model performance and research impact.
