+++ JOSS/paper.md
@@ -162,11 +162,11 @@ Table 1 comes from `examples/mmm_three_way_benchmark.ipynb`: same CSV and split,
 
 Table: Three-way benchmark on `examples/data/MMM Data.csv` (190 DMAs × 109 weeks). National = single national weekly series; Panel = geo × week. Ridge row uses Robyn-style inputs (not Meta's full Robyn unless `robynpy` is enabled). Bayesian runs use modest MCMC budgets.
 
-On this dataset, Meridian reaches the highest train R² in Table 1 but very poor holdout R² and RMSE—a large train–holdout gap under the notebook’s modest MCMC settings; that pattern signals weak out-of-sample fit here, not a universal statement about the library. PyMC-Marketing delivers the strongest holdout R²/RMSE among the panel rows, with the longest execution time. DeepCausalMMM is much faster than PyMC in this run while keeping stable positive holdout performance (~0.84 R²), so it occupies a different point on the speed–accuracy tradeoff for this panel; it does not beat PyMC on raw holdout in the table. The national Ridge row is not comparable to panel rows by R² alone. `examples/dashboard_rmse_optimized.py` gives Training R² ≈ 0.95, Holdout R² ≈ 0.84, ~11 pp gap—aligned with Table 1.
+On this dataset, Meridian reaches the highest train R² in Table 1 but very poor holdout R² and RMSE, a large train–holdout gap under the notebook’s modest MCMC settings; that pattern signals weak out-of-sample fit here, not a universal statement about the library. PyMC-Marketing delivers the strongest holdout R²/RMSE among the panel rows, with the longest execution time. DeepCausalMMM is much faster than PyMC in this run while keeping stable positive holdout performance (~0.84 R²), so it occupies a different point on the speed–accuracy tradeoff for this panel; it does not beat PyMC on raw holdout in the table. The national Ridge row is not comparable to panel rows by R² alone. `examples/dashboard_rmse_optimized.py` gives Training R² ≈ 0.95, Holdout R² ≈ 0.84, ~11 pp gap, aligned with Table 1.
 
 # Research Impact Statement
 
-DeepCausalMMM fills a PyTorch-oriented niche: installable multi-region MMM with holdouts, Hill saturation, and upper-triangular channel coupling, alongside mainstream Bayesian tools (PyMC-Marketing, Meridian). Table 1 and the benchmark notebook provide a shared reference on one public panel (DeepCausalMMM ≈ 0.84 holdout R², ≈ 11 pp train–holdout gap). Near-term significance—with community evidence still early—comes from PyPI, Zenodo DOI (metadata), Read the Docs, and CI for Python 3.9–3.13, which lower the barrier to try, reproduce, and extend the software; citations and course adoption remain limited, and private industry results are not reported. Practitioners on Python DL stacks and researchers needing a reproducible baseline benefit most.
+DeepCausalMMM fills a PyTorch-oriented niche: installable multi-region MMM with holdouts, Hill saturation, and upper-triangular channel coupling, alongside mainstream Bayesian tools (PyMC-Marketing, Meridian). Table 1 and the benchmark notebook provide a shared reference on one public panel (DeepCausalMMM ≈ 0.84 holdout R², ≈ 11 pp train–holdout gap). Near-term significance with community evidence still early, comes from PyPI, Zenodo DOI (metadata), Read the Docs, and CI for Python 3.9–3.13, which lower the barrier to try, reproduce, and extend the software; citations and course adoption remain limited, and private industry results are not reported. Practitioners on Python DL stacks and researchers needing a reproducible baseline benefit most.
 
 # Reproducibility
 
