### 2026-04-25 1fb487e
docs(paper): render Table 1 superscripts via LaTeX math mode
Unicode superscript glyphs (U+2075..U+2077) were silently dropped by the
JOSS PDF font during xelatex rendering, so cells like "1.7 x 10^7" only
showed "1.7 x 10". Switch the scientific-notation cells and the negative
R^2 values to "$1.7 \times 10^{7}$" / "$-12.43$" so xelatex produces a
real superscript and a typeset minus sign.

Made-with: Cursor
