### 2026-04-25 f07d7e1
docs(paper): simplify Table 1 RMSE to millions to fix column overflow
The previous fix used LaTeX math mode ($1.7 \times 10^{7}$) so xelatex
would render the superscripts that the JOSS PDF font was dropping.
However math-mode cells are noticeably wider than their Unicode
equivalents, which pushed pandoc's column-width algorithm to shrink the
leftmost Method column and caused the text to overlap.

Rewrite the column with the unit in the header (Holdout RMSE (M)) and
the values normalized to millions: 17, 0.53, 0.48, 5.1. This conveys
the same information in narrow cells with no font/superscript
[…]
