### 2026-04-25 #9914 @adityapt
Hey @crvernon - I had to make a couple of changes to fix the overlapping issue in the table, but it seems to be good now with the latest article. Also, reduced the boldening of content across the paper. LMK if you see any other issues (or) if you want me to reduce the bold content further.
