### 2026-04-25 #9914 @adityapt
Hey @crvernon - Looks like it got published, however, the citation seems to be parsing out my compound surname as the middle name. How do I ensure my full surname shows up in the citation of the published paper?

**Citation in JOSS:**
Tirumala, A. P., (2026). DeepCausalMMM: A Deep Learning Framework for Marketing Mix Modeling with Causal Structure Learning. Journal of Open Source Software, 11(120), 9914, https://doi.org/10.21105/joss.09914

**Expected Citation**
Puttaparthi Tirumala, A. (2026). DeepCausalMMM: A Deep Learning Framework for Marketing Mix Modeling with Causal Structure Learning. Journal of Open Source Software, 11(120), 9914, https://doi.org/10.21105/joss.09914
