### 2026-04-25 #9914 @danielskatz
@adityapt - see https://joss.readthedocs.io/en/latest/paper.html#name-parts
