### 2026-04-25 f2d6219
docs(paper): use explicit given-names/surname for compound family name
JOSS's auto-citation heuristic was treating the last whitespace-separated
token of "Aditya Puttaparthi Tirumala" as the surname, producing the
incorrect citation form "Tirumala, A. P." in the published Journal of
Open Source Software record (DOI 10.21105/joss.09914).

Per JOSS's paper metadata docs (Names -> Name parts), compound surnames
should be expressed via explicit given-names/surname fields rather than a
single name string. Replace the flat "name:" entry with given-names +
surname so the citation renders as "Puttapar
[…]
