### 2026-04-25 #9914 @adityapt
Thanks @danielskatz !

I have updated `paper.md` to use explicit `given-names` / `surname` per the JOSS docs:

Updated File: https://github.com/adityapt/deepcausalmmm/blob/main/JOSS/paper.md
The corresponding `CITATION.cff` already records the same split.

Could you please regenerate the rendered paper, JATS metadata, and Crossref deposit from the updated source so the citation, indexing, and "Cite this paper" widgets reflect the correct surname?

Thanks very much!
