### 2026-04-25 #9914 @xuanxu
You can not generate a preview PDF once the paper is published. To add the changes to the final paper the reaccept command can be used.
